[roll] Roll fuchsia [bedrock] Teach Dict to TryIntoOpen
Previously the Dict capability did not support conversion into Open. It
turns out with a few lines of logic we can convert a Dict to Open:

- Recursively convert all members to Open
- Put those members in a vfs directory
- Call `dir.clone().open()` in the OpenFn

This sets us up for the next step which is to use it in
serve_processargs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880986
Original-Revision: def5efe970fdd8572474be7c28235b31b575c918
GitOrigin-RevId: 955f07fa398ef27f035ab4bed486f10ab49093d5
Change-Id: Icfac3c973df7f0b917610201684cf5af4e322c3c