[roll] Roll fuchsia [bedrock] Implement Remote for Open and use it to serve the namespace
After converting Dict to Open, now we implement Remote to get a
fuchsia.io/Directory connection.

The meaning of Remote for Open is to open the capability and return a
client endpoint.

The primary advantage of this design is that both Remote and TryIntoOpen
now has a purpose:

- TryIntoOpen tries to convert a capability into Open, which integrates
  with fuchsia.io. This is a place for us to put a variety of fuchsia.io
  logic, such as node name validation.
- Remote convert a capability into a handle. In case of Open
  capabilities, it will always open the capability with semantics
  determined by fuchsia.io

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881351
Original-Revision: d4b037d3ea9525354c8df30862603d57fd43a502
GitOrigin-RevId: ff3ed5856fa6c552694859a86e3531a8c623ac23
Change-Id: I82a40423c1489f02dd3bcec831398673f4233ed9