[roll] Roll fuchsia [inspect] Route InspectSink to connectivity
This is in service of RFC 168:
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0168_exposing_inspect_through_inspectsink

InspectSink will replace the out/diagnostics directory.
For an easier transition, InspectSink is routed by
the diagnostics team before a compiler flag is added
to cmc enforcing the offering of InspectSink.

These changes are primarily automated and either add
manual routing of InspectSink to children and collections
in shards, or add the inspect/client.shard.cml if the
CML represents a final component (ie non-shard).

The change touches so many files as the offer shard uses `offer InspectSink to "all"` syntax, which is viral when in included in shards. Anything that includes one of the connectivity shards at any depth will end up with an offer of InspectSink from parent to each child and collection.

As a general rule, that would be an SDK compatibility issue for OOT users (eg with RealmBuilder shards), and make it impossible to include many shards in CMLs that need to do custom InspectSink routing (eg in some Archivist stuff), so the automated tooling did not insert offer shards into other shards.

Original-Bug: 93344
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/881452
Original-Revision: 2f8aa33e1242a540adda2770d6407d1ea5d567ad
GitOrigin-RevId: 2186d2b805877966730d247ac940b23d4fd270f0
Change-Id: I80e6f19fe5d6f8959fe1b20b8fb63d6d91e1f2c7