[roll] Roll fuchsia [fdf][ddk] Add zx_device_t* parameter to get_root_resource
This allows us to lazily fetch the root resource and avoid using a
global to store the cached version of it in case the driver requests it
multiple times. This is overall more efficient and allows us to avoid CF
logging warnings every time a failure occurs due to a proactive request
that inevitably fails.

A new global was added to the x86 acpi board driver to route the root
resource to code which requires it. Piping it through the acpica code
would make it more challenging to keep up with upstream whenever we sync
to newer revisions of the library.

Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862216
Original-Revision: 4a414ccb794faf8fcd90333cd9a4b91d02f83757
GitOrigin-RevId: ccc80d6d3980d9904e5f13eeb9851a8f188ad135
Change-Id: I6c88464147527a79b72411fc12e878499b0265d4