[roll] Roll fuchsia [driver_runtime] Correctly handle no signal packet
The signal packet may be nullptr when the wait is canceled.
Dereferencing it causes a crash, so instead we store a std::optional<>
signal packet to remember if we recieved a valid signal packet to begin
with.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883812
Original-Revision: 9eef1dd80f4e10ba5cc3a4ce978a79bed1d4bc26
GitOrigin-RevId: 3e300bf26c0c77bb65b083e7e1e98137ebfea526
Change-Id: I3e760f2426b998e6c983dad0ce9e9cfff75f2141