[roll] Roll fuchsia [installer] transfer ownership to firmware
None of the current installer owners work on Fuchsia anymore; for the
time being transfer ownership to the firmware team.

I was planning on using the higher-level OWNERS for a while longer, but
it turns out there are a handful of other projects that include this
OWNERS file, and that type of indirect ownership does not fall back to
parent directories, leaving those other projects essentially ownerless:
* //build/images/installer
* //src/sys/live_usb
* //src/sys/lib/payload_streamer

So we might as well just transfer ownership now.

Original-Bug: b/282797139
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884051
Original-Revision: 7f2b4bba4e76d04fade9804ef555d4ba6b61196b
GitOrigin-RevId: 0088b8666081aa716a99b702de2a9fbe0784c970
Change-Id: Ic75e151e1b38e1fda9e971b168a670475e1fd9e3