[roll] Roll fuchsia [netstack3] Handle unbound in dgram listen_inner
Make socket::datagram::listen_inner take a SocketId instead of a
listener socket ID and return an error when that socket is already bound
from listen_inner instead of switching on the socket ID type outside.
Since the socket ID type is going to soon be un-differentiated, this
check needs to be done by looking at the state held for the socket, not
at the socket ID itself.

Original-Bug: 126141
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883931
Original-Revision: f37958432bdb6771b66797481a9506fd39809083
GitOrigin-RevId: 93391ff5ba518278340f2f2b7d04846d4d735b69
Change-Id: I2eb78f6dcdfa9414ab5d90ce280625e38eeec6af