[roll] Roll fuchsia [rbe] Suppress writing stub on matching output
This accounts for interactions between the rewrapper features
--download_outputs=false and --preserve_unchanged_output_mtime.

When a download is skipped, it could be attributed to either
of these features, but download stub creation should depend
on which case it is.

If an output already exists and is valid (its digest matches
that reported by the corresponding output_file in the action_log),
we want to leave it as-is, without overwriting it with a download stub.
If the existing output is a download stub, we update it only if the
blob_digest changes.

This lets us continue to have the benefit of leaving timestamps
untouched where it is correct to do so (for ninja restat).

Original-Bug: 123178
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864305
Original-Revision: 39db2edcdd531c890f4eefbe3e4838204178573c
GitOrigin-RevId: 3cdf55de187b62aabeb5961ba3f12a740f0d705c
Change-Id: I7d99caf0d4c4afc9144598a1442772f1d52986af