[roll] Roll fuchsia [elfldltl] Do not zero beyond the VMO size
Fix a bug in VmarLoader::MapWritable which would zero past the last page
of the VMO if the size was page-aligned. Without this fix, the
elf-search test fails with fxr/883224 that enforces out-of-bounds check
on the specified range instead of trimming it to fit within the VMO
size.

Original-Bug: 130470
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883542
Original-Revision: e73d1c798bc3c4312bbb25aa99e6f42d8317e672
GitOrigin-RevId: 1b00603a06dee543a6e83fbdc40ccfde2b2c8e03
Change-Id: If7a62d1e70883cb0b1d9aaa44433d6a2d105f5bf