[roll] Roll fuchsia [assembly] Move default mdns config to assembly
The products are still adding private mdns configs using GN, but those
configs are added to a separate destination and merged at runtime. See
the bug for more details.

An AIB is added to add the mdns service for wired udp so that the device
can be discovered by ffx.

This AIB is not enabled on user builds. It is added by default on
userdebug/eng builds, but can be disabled explicitly by the product
config.

Original-Bug: b/290984622
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883632
Original-Revision: 97102ef6f57ea8e78d052535615f9328082ef311
GitOrigin-RevId: d8cea55777716ac1449a61bddee4ef5f50999d26
Change-Id: I3691a496a09c760476fc1c5c31a12d2f8cc8c3b2