[roll] Roll fuchsia [display][banjo] Rename "Client" to match FIDL type and change type to bitfield.
The "Client" type in display banjo API (corresponding to
"ClientCompositionOpcode" in display FIDL API) has a typename that
could be confused with other display coordinator / driver concepts.
Also it should be a bitfield instead of an enum.

This change renames the type to "ClientCompositionOpcode" and fixes
the type expression in banjo API.

Also CheckConfiguration() should return an array of "Client" bitfields
instead of array of uint32. This changes the return value type as well.

Test: no binary change.
Original-Bug: 130426
Original-Bug: 130427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882740
Original-Revision: 96c7ddec91ca4506ab7d96efd1f0428609e06799
GitOrigin-RevId: b743960db115b7d6e84aab24a72c24f588444d59
Change-Id: I7c34c9de00ef61865d344ea72cba5b1b44adba0a