[roll] Roll fuchsia [scenic][flake] Capture a weak ptr instead of this
It seems that capturing `this` in this specific closure can make the
this pointer outlive the actual object.

The simplest way to remove that issue is to use a weak ptr instead,
but it is not super-obvious that no other issues lurk here.
Specifically, `Endpoint` destruction is nontrivial.
It still seems like an adequate thing to do.

Original-Bug: 128914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883533
Original-Revision: f4931e2b4c290ccac55bf09cb259e3bb24b5320d
GitOrigin-RevId: 669a0545b0b5e1e0b9b475a8312f41a640d3fb0d
Change-Id: I7724c9cbcd7e7c0f037b90ef2367f4ad45afd3c0