[roll] Roll fuchsia [fdf] Don't return host koid when unbound
Currently when the node is unbound we can still return a host koid. This
is because the driver host pointer being used is a raw pointer into the
self deleting driver host list entry. It doesn't make sense to have a
host koid for an unbound driver, so now we check to make sure the node
is bound before trying to find the host koid.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883717
Original-Revision: 437dd42b51545f19d51c95811005e1aa807402c2
GitOrigin-RevId: dc8503aca58a829c8af5cce6e7967674edcf5fb7
Change-Id: I04a2c163f5b62c9b5e4c9c40f1a0ab860d6e3fee