[roll] Roll fuchsia [async-loop] Fix race between PostTask and Shutdown
If a PostTask was running and Shutdown is called at the same time, the
PostTask could succeed in inserting its task into the list after the
shutdown finished. If that happens, then the task will never be run (not
even with ZX_ERR_CANCELED) which could cause problems. For example, with
an async::PostTask the memory allocated for the task will be leaked.

Multiply: async-loop-test-pkg
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884075
Original-Revision: adf033c790aeaf10eb311d2dc8b03a6f53159721
GitOrigin-RevId: 8e625e47a6ec49df6518a9856ca07c4cd30f23a5
Change-Id: Ib794a8ea11ed2b7f96078d7bc0941d8cf5b02cad