[roll] Roll fuchsia [starnix] Store loop device offset in device state.
The offset is currently stored in the file object and since the test
included in this CL fails at ToT that doesn't seem to be synchronized
across all file descriptors.

This change moves the offset state into the loop device itself so that
the configuration will be visible to all users of the loop device.

Original-Bug: 129705

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882573
Original-Revision: b603bfa0d3f4004e927cf0a8d9ae4ab1b041f7e4
GitOrigin-RevId: 0b4e7fdd44e1c9aa74d8b63a598f2b466a5d4d4b
Change-Id: I82ef9a9353618cccaec787608c5f452f20d9d7a6