[roll] Roll fuchsia [ui][input-pipeline] remove unused `Clone` from LightSensorHandler
`LightSensorHandler::new()` returns an `Rc<Self>`. Hence:
- every instance of `LightSensorHandler` created by `new()`
  implements `Clone`, [by virtue of `Rc`](https://doc.rust-lang.org/std/rc/struct.Rc.html#impl-Clone-for-Rc%3CT%3E)
- there's no need for `LightSensorHandler` to implement
  `Clone` itself

The primary benefit of having `Clone` come from `Rc`, is that we
know that all references are functionally identical, without having
to look inside the definition of the `LightSensorHandler` struct to
verify that all of the mutable data is shared across instances.

A secondary benefit is that we can remove `Rc` from the fields of
the struct, and thereby reduce the number of places we might have
to look to find cycles in the object graph (if we run into memory
bloat).

Original-Bug: None
Test: fx test input_pipeline_lib_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884200
Original-Revision: 488985e7db43107a21e8d39d4bb0f074bcd61622
GitOrigin-RevId: 478b5a3e6e54ee2a0281564ef1a643337423c90d
Change-Id: I88889a1ba9a1f92bcd0c5ccd0e2eadad6940b536