[roll] Roll fuchsia [ld] Refactor bootstrapping
This refactors bootstrapping so that the ld::abi Module data
structures for the vDSO and for the dynamic linker itself are
filled out along the way.

It also cleans up the build details a bit, renaming all the
targets to be called ld-* instead of *-ld. The ld-startup target
is now a shared_library, so it will use the user.basic variant
corresponding to the variant of its dependent (test).  The files
like this used by the test but not linked into the test program
go in data_deps, not deps. This matters when they include
linkable targets like a shared_library.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883513
Original-Revision: c344b57ef72bab33c99d9b38bac598063affcd04
GitOrigin-RevId: 15f1aae5e490dcbf21cd18008d0c4f7458333279
Change-Id: I28421e9699ccb4540b194e37919e5626469da45b