[roll] Roll fuchsia [ffx][fastboot] reland boot timeout success
(This is a reland of fxr/882214 -- the problem was that a
usb_bulk::read() will always timeout after 800ms; when running
a read with a longer timeout, the code needs to retry.)

When issuing a "boot", "continue", "reboot", or "reboot-bootloader",
it's possible that the "okay" response gets lost in a buffer. So inside
of consdering a lack of an "okay" in those cases as a a timeout failure,
consider them successes.

Also, since we are now considering a timeout to be an expected
result, let's timeout in only a few seconds instead of waiting
30 seconds.

This change includes a fix to fastboot::read_with_timeout() to
actually use a timeout, rather than blocking and checking the
timestamp after the fact.  Note that since usb_bulk will timeout
after 800ms, we have to retry if we get ETIMEDOUT.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883118
Original-Revision: 30467cc016a99dbebcdeba124e5679391c05f081
GitOrigin-RevId: e737bed1c3248de6dabf2c8d536da4ef76635581
Change-Id: I6637a337d07706c595cbc27cbe05a2498232604e