[roll] Roll fuchsia [gn][sdk] Use the prebuilt name in sdk_pkg
This fixes an issue with sdk shared_libraries where the name of the
shared library .so was different from the sdk atom. This is the case
when the sdk_shared_library has explicitly stated a sdk_name that is
different from the gn target_name. This change forwards the .so name
into the generated fuchsia_sdk_pkg from the meta.json.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883580
Original-Revision: ab1b7386f58b9267373ddb208ffecb9c85a2a134
GitOrigin-RevId: 5fedb79d676281a90eabc145973f8202e535e086
Change-Id: Iac0f414901706ec9e93c168cd28b5c51813da509