[roll] Roll fuchsia [bt][obex] Create parser and builder for Folder-Listing object
Folder-Listing object is defined in the Obex Specification. It is
expressed as an application of XML. Define parser and builder for it so
that the object could be easily created from and converted to raw data.

Original-Bug: 128458
Test: fx test -v -o bt-obex-objects-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/865636
Original-Revision: d6e8f3fc522f22dcd67f1f84a2f728c6d31132f8
GitOrigin-RevId: e6b663a60ada11382b13242c37f94b7ecdbd9515
Change-Id: Ia02fa27dbeaa0dcb5822e86aa6ec97cba33a014a