[roll] Roll fuchsia [magma] Don't use shared_ptr with MagmaSystemDevice
When creating a device, the driver framework gives the device temporary
ownership of certain hardware resources (e.g. interrupts). Those
resources must released by the driver before driver teardown completes.

When MagmaSystemDevice is held as a shared_ptr by the
MagmaSystemConnection, the lock on it in the MagmaSystemConnection
destructor may hold on to it after MagmaSystemDevice::Shutdown() is
called, which may cause the resources to not be released in a timely
manner.

Since the MagmaSystemDevice always outlives the MagmaSystemConnection
anyway (because Shutdown() tears down the connection), we don't need the
weak_ptr in MagmaSystemConnection; removing it fixes the lifetime issue.

Verifying that the MagmaSystemConnection outlives the MagmaSystemDevice
is difficult with the current way we use shared_ptrs with
PrimaryFidlServer. To fix that and also ensure that the FIDL bindings
are being used in a thread-safe way, add a PrimaryFidlServerHolder class
that's responsible for ensuring that the PrimaryFidlServer is torn down
on the correct thread; PrimaryFidlServerHolder can have external
references taken to it and can have an arbitrarily-long lifetime.

Multiply: magma_sys_driver_test
Original-Bug: 129953
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883133
Original-Revision: 936dae30731acb6ac3067b3cd737dcf7beba0eb4
GitOrigin-RevId: 74261ce38c24d4c4199885cf376aa6d56ab2226c
Change-Id: Ibf1ddc8839a5a115ff7e179a44afb001e9eb8498