[roll] Roll fuchsia [bt][obex] Send & Receive the Connection ID
The Connection ID is used to uniquely identify an OBEX session between a
Client and Server. It is optionally determined in the CONNECT operation,
and if provided by the Server, is used for the lifetime of the session.

The ID is uniquely generated by the OBEX Server and must be included in
the first packet of every subsequent operation after the CONNECT.

Original-Bug: 128855
Test: fx test bt-obex-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871292
Original-Revision: 938db9a362904d17e863f7796db1be9afc3cbdfb
GitOrigin-RevId: acc7d1998d781e02ff8d0d9f95bb227fb865fa03
Change-Id: I42cc15ce2c2c2376d06dc61a66fcfeea324cbd8c