[roll] Roll fuchsia [ld] Plumb processargs protocol
This makes the Fuchsia dynamic linker startup interpret the
actual (legacy) processargs protocol.  It uses this to wire up a
debuglog or socket (from fd 2 in the message) for logging.

This includes preparing two new pieces of test harness.  First,
TestLogSocket is a simple collector for the socket given to a
test for its diagnostic logging, so it can be matched against
expectations.  Second, TestProcessArgs builds and sends the
processargs message, letting the test precisely control the
details of what goes into it.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884247
Original-Revision: 13812748d44542fab9671f8ecad56ea056af13d0
GitOrigin-RevId: dc5a1417951ebc57e4019d00b423a4a6a8d1ba30
Change-Id: I71e17e3e225d0c08aea7d6e7c82097add0931077