[roll] Roll fuchsia [ld] Wire up allocators
This sets up the //src/lib/trivial-allocator machinery for use in
the dynamic linker.  There is a scratch allocator for ephemeral
uses during startup, and there is an initial-exec allocator that
will be used to populate the passive ABI.

For the Fuchsia incarnation's tests, this now plumbs the "root"
VMAR through processargs.  The in-process tests use a 1GiB VMAR
to isolate allocations, and destroy it at the end of the test.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/885071
Original-Revision: 0e378c66d515b35f21d115decc79e6b5578e97f1
GitOrigin-RevId: de935e556a15405c5df957aca9721eb38f4ee4e6
Change-Id: I0092f6d71cb8879a327bcd3353d64fb69fd176f8