[roll] Roll fuchsia [bt][obex] Encode headers in order
Use a BTreeMap to store headers in the HeaderSet collection. This
provides for efficient insertion and removal, and fairly efficient
lookup. The size of this collection is bounded by the number of unique
HeaderIdentifiers (26 + user defined). Prefer BTreeMap over HashMap due
to the ordering invariant that exists for the BTreeMap. By maintaining
an (arbitrary) ordering, verifying correctness in unittests is
simplified as the encoded buffer will be encoded in the same order each
time. With a HashMap, encoding ordering is not guaranteed and makes
asserting data validity difficult.

Add TODOs for future ordering constraints when the relevant features are
added.

Original-Bug: 125070
Test: fx test bt-obex-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872722
Original-Revision: 4ccfa1cb0769bd4ca19a901929153a72ab3ea8f1
GitOrigin-RevId: d21829c1275212966785eef7ed536b72933cfacd
Change-Id: I5947f07efc7ef4010f4df5ecf5396a591d9fef0a