[roll] Roll fuchsia [emulator_test] Disallow global owners
This test is designed to block large scale changes in fuchsia.git that
might break the Netstack team's out of tree conformance testing.
Disallow global owners on the test source so that a Netstack team member
will have to be included when the test needs to be revised (therefore
signaling a potential breakage).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884159
Original-Revision: 307bae93fbeebaa15e232b776d7693ab0908501a
GitOrigin-RevId: ebc2159479c7528df143f0796c864f8f29a27737
Change-Id: Ica73f306cc47c1632647f2548deb0bce9f93ed20