[roll] Roll fuchsia [socket-tun] Send message with varied bytes
As a result of discovering and fixing https://fxbug.dev/128850, change
the bytes that are sent in this benchmark to something more varied than
a single byte repeated ad infinitum, because any permutation of such a
message is equal to the original message and a sender asserting that
the received bytes is equal to the sent bytes cannot tell that the
message was in fact sent out-of-order.

Add assertions that sent bytes is equal to received bytes.

Always set the socket option to modify the send buffer size (this was
not previously done because lowering the send buffer size to 10KiB was
what tickled the issue).

Test: Manually verified that tun-socket-benchmarks-ns3-with-tracing
      produces traces that do not exhibit the symptoms fixed in
      https://fxbug.dev/128850.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874976
Original-Revision: 41e3239f69aff14d90f8c24a5f628a31394d8cc5
GitOrigin-RevId: 6df1b6561abbc6d9a1c23266e81ce0fba25bb792
Change-Id: I26f2c574e0f47b920b17d895719818c63959898b