[roll] Roll fuchsia [starnix] Loosen what waitpid test records
The test involves checking to see that a SIGUSR1 was
appropriately delivered when the child exited. We see a flake that
looks like the handler got more than one signal, and that
one of the signals was a SIGCHLD. I hypothesize that SIGCHLD
happened because of a suspend, rather than an exit.  This CL
ignores signals that happen for a reason other than the child
exiting.

Original-Bug: 130640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884701
Original-Revision: a550ec17bb85da16ddc25268707fe933c8071b19
GitOrigin-RevId: 066087b1b602de32d31f851b7e2fc8b22ed52df5
Change-Id: I8746a8e416316f03f3d52c339747a504508b93d6