[roll] Roll fuchsia [reland][assembly] Introduce bootfs-drivers GN logic
This is a reland of fxrev.dev/869283.

Introduce `boot_driver_packages` GN logic that supports a list of
boot driver packages that should be included in the boot driver
manifest and packaged into Bootfs.

The `--boot-driver-pkg-list`/`--boot_drivers` arguments are plumbed
through AIB tools to Assembly Builder to parse and generate a
corresponding driver manifest and include in the BootFS package set.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886571
Original-Revision: 93830c6c4958494b85b6673766f8ab947915de53
GitOrigin-RevId: a972d078cf8ee057851c4cf6258b1dd5fadf2af0
Change-Id: I1e88c8ccd0ab320a44e3b3384cdb154ae7ea38e3