[roll] Roll fuchsia [scenic][tests] refactor BlockingPresent for flatland
Refactor the way that BlockingPresent() is implemented
for `flatland_integration_tests`, so that there's only
one implementation of the function. (All of of the
existing implementations were identical.)

This is in preparation for changing the function to
also block until the next frame can begin.

Note that this isn't strictly a refactor, as the
migration of code changes the log messages about
where `RunLoopUntil()` was called from.

Leaving the arguments to `BlockingPresent()` unchanged
would have mean that the log messages around
`RunLoopUntil()` would reference `blocking_present.cc`,
rather than the specific integration test fixture's
source file.

Instead, we now pass the source location through
from a `BlockingPresent()` call-site to `RunLoopUntil()`.

Note also: we make `LoggingEventLoop::RunLoopUntil()`
public, so that it can called from the new version
of `BlockingPresent()`, which is a free function.

I'm not really sure why those methods are protected
in `loop_fixture::RealLoop` to begin with, TBH.

Original-Bug: 130653
Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884795
Original-Revision: 5ab0861de3b3bdeb176572e57963d0ad908a7223
GitOrigin-RevId: 63afb9b06c134dcca449534bdd01ee5236c35e73
Change-Id: Ie4cd375b0c7fdeed8c553a8d9372643bb5e42e15