[roll] Roll fuchsia [starnix] Split filesystem creation between CurrentTask and Kernel
Filesystems can be created from a Kernel object and an external resource
or they can be specific to the currently executing task. FUSE
filesystems and ext4 filesystems backed by open files are examples of
the latter. This splits the filesystem creation routines based on the
caller's context so that filesystems that are not task specific can be
created without a CurrentTask reference.

Original-Bug: 130182

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/885633
Original-Revision: b2bec9caa00cb15a06acd8ed91ea5938377e7d02
GitOrigin-RevId: 14c6289a8966e434de40f1b60eeb865616ba94bf
Change-Id: I7a848764b56e1e40e5a193b16f7f12ee507c25d6