[roll] Roll fuchsia [fake-display] Add CaptureImageInfo class; Use DriverCaptureImageId to identify capture images.
FakeDisplay used to use ImageInfo for both display images and capture
images, and refers to imported images using raw pointers. This is
prone to use-after-free errors and could cause images of incorrect
type to be used.

This change adds a new class CaptureImageInfo to store metadata for
capture images. Also it adds an `id_` field with type
DriverCaptureImageId to identify the image, which will be the value
of returned capture_handle in ImportImageForCapture().

This change doesn't change any invariants (explicit or implicit) on
validity of imported images.

Test: fake-display-tests
Original-Bug: 128706
Original-Bug: 128653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882471
Original-Revision: 8fcca2215bb0fa14f75774648b354d8bbaf9d16f
GitOrigin-RevId: 0d82533fcc254097bb7b61559ce4b24d581485cc
Change-Id: I7c5262be750df577562130e3ca445e737255ab80