[roll] Roll fuchsia [fxfs] Add dirent caching to LSM tree.
Wraps both dirent objects to reduce the memory from the full
ObjectValue enum. This required changing the type signature, but
won't affect the actual serialization.

The cache lives on top of the whole tree, which means that it may
duplicate items in the mutable layer but also that there is no
resolution required when sealing.

The cache is only used for volumes, other lsm tree instances in Fxfs are
given a `NullCache` that essentially does nothing.

Original-Bug: 128188

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875681
Original-Revision: 06bd7df233d1cf69df5e8ec4069b971840698857
GitOrigin-RevId: 99a6bfd0ff2ff771836a6a76b5d570c1c0f46969
Change-Id: I63f2f9352fd07ef2861cc57954eec44bf32c1afe