[roll] Roll fuchsia [bt][hfp] Move call types to a library.
This change move the call types that we expect to use in the HFP
Hands-Free component out of the HFP Audio Gateway and into a library to
be used by both components.  We expect to add more types to the library
crate, bt-hfp, as HFP Hands-Free is developed, and the name reflects
that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872540
Original-Revision: 7353ee66563189928abebeea042d2cedf9025dee
GitOrigin-RevId: 4435d5a0ad1753341572de9d9b3ecc106880fd2f
Change-Id: I723e3f10d1b0b7baa07f20415b532bb7eecddcc9