[roll] Roll fuchsia [perfcompare] Handle bigger-is-better units
Before this change, perfcompare would treat all metrics as
smaller-is-better.  Fix that so that metrics may be treated as
bigger-is-better based on their units, matching the treatment in
Chromeperf.

We now report metrics as "improved" or "regressed" rather than
"faster" or "slower".

Original-Fixed: 111103
Test: perfcompare_test.py
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884333
Original-Revision: a29cf60bce6fac6702f1cbaba48e192ae3691c80
GitOrigin-RevId: cae4fd8a1d984706a209600ab320b25625c2e2fc
Change-Id: I1bd2217c6cd1cb58d19ec99ebb00f9a7251666e6