[roll] Roll fuchsia [ld] Partially plumb llvm-profdata in startup dynamic linker
This sets up the machinery for the instrumented builds of the
startup dynamic linker to collect their llvm-profdata into a VMO
for publication via the debugdata protocol.  There isn't yet a
way to get that VMO anywhere, so it's just dropped on the floor;
fxbug.dev/130542 tracks actually getting the VMO published.

Unfortunately, this doesn't entirely work as intended because
live function invocations continue to use a stale pointer into
the data segment location of the instrumentation counters.  This
means the tails of functions still running after the counters are
redirected to the debugdata VMO will have their updates lost, so
in future we may need to reorganize the code to avoid that
problem.  For now, the more pressing problem is that writing via
the stale pointer after protecting the data segment will get a
page fault.  So for now protecting the data segment is disabled
in the builds using llvm-profdata instrumentation.

Original-Bug: 130542
Original-Fixed: 130542
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886591
Original-Revision: 9f1b3e21cf6e0211d1c9ab78c2880a23f17f463a
GitOrigin-RevId: 43cbd56807a679e311d00b4fe7a217943baea2c4
Change-Id: I1a2b61c5f27384c9c796b73278cd4845f5be59d3