[roll] Roll fuchsia [fxblob] BlobDirectory::lookup now checks for CREATE before WRITABLE
This change is in anticipation of removing support for the old
fuchsia.io write API in fxblob. Pkg-cache relies on create_blob()
returning ALREADY_EXISTS to know whether or not to write a blob so when
create_blob() calls lookup with the CREATE and WRITABLE flags, we want
lookup to fail with AlreadyExists and not AccessDenied from the check on
the WRITABLE flag.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886711
Original-Revision: 91dc0bc0ca059dd500013b4a06227177f31d776f
GitOrigin-RevId: df96f2d2783df0735d4b95b0205936165985cb5c
Change-Id: Ie34b21d06813ee1b050d0ac3a2c31f66365a99e4