[roll] Roll fuchsia [boot-shim] Introduce CPU topology decoding.
While specific for RISC-V, most of the complexities are shared with ARM
and will be generalized once RISCV support is complete.

This introduces an item capable of generating the CPU topology data from
the devicetree.

For performance class, all clusters have a default performance class of
1, unless a CPU specifies scaling property wihin that cluster. At that
point the highest capacity is picked as a representative of the cluster.

Then the performance is scaled down to 0-255 range where 255 is the
performance class of the cluster with the highest performance class.

Performance class is meant to be a relative comparison between CPUs, for
scheduling purposes.

Also the devicetree boot shim now appends the CPU topology for riscv.

Original-Bug: 114124
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879312
Original-Revision: 4829500a344f9a0e2f70c5aaa7e7dfc6518b52fb
GitOrigin-RevId: a0aaec8ca0362e85bf7d2893426fac3efce61140
Change-Id: I0359656ccb1c01a7df815e9a59e2421e9cceabce