[roll] Roll fuchsia [amlogic_decoder] Implement driver release hook
This was historically not necessary to implement because we never call
the releas hook in DFv1. In DFv2, the compat driver will actually call
it when the number of driver instances for a particular driver hits 0.
This happens during the amlogic decoder integration tests which unload
the driver in favor of another test driver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887229
Original-Revision: 049d481d0b933b41e0c412719ce3691b61f3557e
GitOrigin-RevId: f201b87f86fa0dffec768a9ccfd8480ac96077c5
Change-Id: I3391f10020eda6a71da1c2b9b0298ceeeadb4d55