[roll] Roll fuchsia [vm][test] Rewrite Pager.VmarRemapTest
This test overwrites a VM mapping from one pager-backed VMO to another,
and verifies that blocked threads see the new contents when unblocked.
Rewrite the test to use a mapping that the test itself creates, instead
of relying on the one created by the Vmo instance. This avoids using
UserPager::ReplaceVmo().

Remove UserPager::ReplaceVmo() as this was the only test using it.
ReplaceVmo() would replace Vmo::vmo_ under the Vmo's lock, but a caller
might have a const reference to the vmo_ outstanding at the time, which
could lead to data races.

Test: core-pager
Original-Bug: 129315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886413
Original-Revision: 208fa844729f5988842d272b9945794fadbd0dab
GitOrigin-RevId: a8f32a33a5cfdec9d69b799f4908f0ba34c87394
Change-Id: Iffa860ac5214f4edd225eb11ea3ea440d02ff2a6