[roll] Roll fuchsia [ffx][config] Add flag to ffx for config domain root
Implements `ffx --env-root <path>` which points at a path to search for
a configuration domain root in. It also adds support for using a
configuration domain with an isolation root, though the exact parameters
of that are subject to change.

`--env-root` is currently hidden as the interface is subject to change,
and this is being implemented mostly for dogfooding to find the right
tradeoffs here. The flag may be merged into `--env` later.

Original-Bug: b/290946533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887227
Original-Revision: 13a9059e737bc4a77413d7d59b1bee4e3dfda6b2
GitOrigin-RevId: 034ecd0a8501f18d51036c9cd0073d12ad511971
Change-Id: Ic4d07fd130641ebbca814050f7ee2ca6dba9fab5