[roll] Roll fuchsia [gigaboot] Add managed network ethernet agent
Add an ergonomic wrapper around the managed network protocol in the
form of an EthernetAgent.
The agent is responsible for
*) Finding a valid network interface
*) Constructing the layer 2 header
*) Transmitting a payload

The agent is not yet used by any other code, but it is intended to be
used by the mDNS broadcast rewrite.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886653
Original-Revision: d62d2348df602bd255a94008477b42bf1bd55725
GitOrigin-RevId: fe38d2c2d2a64c177498dc3bf1ba293cfb55b2e4
Change-Id: If0573335ea97f79d41571602bdbf6a8735c53b73