[roll] Roll fuchsia [assembly] BoardInformation is no longer optional
Require BoardInformation to be specified for all assembled systems.

This required also ensuring that any host-test that runs assembly, and
uses its output as test data, isn't included in the build when a board
is not defined (sdk builders, and host-test-only builders are in this
class of configurations).

Also moves scrutiny tests into the tree of 'host_tests' groups, and
only defines those tests when a board has been defined for the build.

Original-Bug: b/287519036

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886659
Original-Revision: 3b0d6bfff15a1107cd579e33e4fafcd35cc74d40
GitOrigin-RevId: 2dcde20afff5bbf6adb1891310e2f657cc586801
Change-Id: I959fff8f12dd19696f4e4839184be2f8b74d314b