[roll] Roll antlion Lazy initialize wlan_policy_controller
Fix an issue where SSH and SL4F was not being verified after a hard
reboot due to a failed test. This was due to WlanPolicyController
storing its own reference to the FuchsiaDevice's ssh and sl4f
properties. When the device is rebooted, the FuchsiaDevice's references
are deleted and recreated. This however did not delete or recreate the
references stored within WlanPolicyController. This caused Python not to
garbage collect the old ssh and sl4f modules, thus bypassing the initial
connectivity checks in their __init__ functions when used by
WlanPolicyController.

Original-Fixed: b/292146900
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/887874
Original-Revision: df3101bc6c797ae905336b469985c6099128af94
GitOrigin-RevId: 51703ae28c175ef1816c30ea0a34f05eddf92838
Change-Id: Iffd26e9df3c60148a4fa5e72c410933e036ae3da