[roll] Roll fuchsia [assembly] Deprecate fvm_emmc/fvm_ftl gn args
These two args:
- fvm_emmc_partition_size
- fvm_ftl_nand_block_count

Are only used in assembled_system() to determine whether to declare the
optional fastboot and nand fvms as outputs. We are replacing these two
arguments with ones that are better named.

We are also using this opportunity to remove the usage of a few global
gn arguments in assembled_system() and forcing clients to specify the
images they expect to be generated:
- generate_fvm
- generate_fvm_fastboot
- generate_fvm_nand
- generate_fxfs
- generate_vbmeta

Original-Bug: b/291945693
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887234
Original-Revision: 31b4a87353ebc159765391f79aae86e83ce455de
GitOrigin-RevId: aae7d1276b36dc767831c087b51ec0a5c17a7aa8
Change-Id: Idecf520b2a44525dc363e3b2b5abd03aa0df711c