[roll] Roll fuchsia [rust][vfs] provide path, not name, to not found handler
A simple directory in the Rust VFS may have a handler registered with
it, which is called whenever an open request is received for a path that
the VFS is not configured to handle.

That handler should be called with the full path that cannot be served,
instead of the single node in the path that the not found issue was
caused by.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879904
Original-Revision: 4b24b2741cb2d0616d5c551510f61307ed2cd3cb
GitOrigin-RevId: afa64aa101c24019ef7b89c787b763e6d61e851b
Change-Id: I7c18b3be031dadcf5d11b1c05607895d4c9bdde7