[roll] Roll fuchsia [netlink] Synchronize route sink multicast
Refactor FakeSenderSink to be backed by a Sender/Receiver pair instead
of a vector. This allows clients of FakeSenderSink to fetch all
available messages, even if they are not available immediately, as long
as they will be ready.

Thanks to @martinjeffrey for helping to come up with a synchronous
approach to this problem so all the callsites of `take_messages` did not
require an update.

Fixed a race in the RTM_NEWROUTE test that occurred when attempting to
check for a new message immediately after adding a new route. The test
now blocks while waiting for the expected messages to arrive.

Original-Fixed: 130548
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884054
Original-Revision: 3e7941932d64b376755d7dcb59b0f59704a20cd1
GitOrigin-RevId: a621a843c99473a0dda47d1d100bdd90fa30beae
Change-Id: I08b801d15d57dd926a633919d028acce6ced5dda