[roll] Roll fuchsia [storage][watchdog] Allow longer timeouts on watchdog tests.
Occasionally we get flakes on the watchdog tests since it sets a timeout
of 1 second then waits 2 seconds to see if the expiry call ran. This
adds a looping poller to give up to 10 seconds for the expiry call to
schedule and run rather than just 1, but polls every 100ms to keep the
runtime of the test as low as possible.

Original-Bug: 130876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888231
Original-Revision: c6c3e9ffd84883e6bdb62ff797e8ce4a0267eb7e
GitOrigin-RevId: 7f7d77075da0867240759d9c2cb3a49b050b1ed7
Change-Id: I1463197c22f9897929aa08726fa06448b98a12dd