[roll] Roll fuchsia [vfs][journal] Fail replay on incorrect payload prefix
Previously this would cause a runtime assertion which would crash Minfs.
Now we detect this condition and stop replaying the journal.

This should already be guarded against in production by the entry
checksum, which is verified before decoding the entries. However, if a
bit flipped before we calculate the checksum, we may persist a journal
entry with this condition. This was found by the fuzzer, which skips
verifying the checksums.

Original-Fixed: 130131
Test: fx test journal-unit-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888331
Original-Revision: 24e190f7b08696d3604800ff1f951e59483cdeac
GitOrigin-RevId: bdf2909590623c94d53c40ddb190dbe16f9c0722
Change-Id: Ib023616e03e05a827e994f87a7789bb2f375fa3b