[roll] Roll fuchsia [starnix] Redirect build container package to the default toolchain
When an android_starnix_container template is instantiated under some
toolchain that is not the default, that would result in gn gen errors
as it tries to reference dependency packages under the wrong toolchain.

We already have some uses of this template but it just so happened that
none of the .gn files instantiating the template was referenced from an
alternative toolchain so it appeared to work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887976
Original-Revision: c416454102875a5e77a097c8a7f9cb3bc26db87d
GitOrigin-RevId: 2b7fd58ed7e5fb81b7cd5ccc3e3e2a897dc9e48d
Change-Id: I23e6cfb80e43eed39e4bac8d8768357897319c06