[roll] Roll fuchsia [starnix] Retry reading coredump report on partial nodes.
The process for creating a coredump report is roughly:

1. create an inspect node in the bounded list implementation
2. record the pid
3. record the argv

The kernel's inspect output is complete before (1) and after (3),
but any reads in the middle will see a coredump report which is missing
one or more properties.

Ideally we'd be able to add nodes to the bounded list only once they've
been fully populated but that would require API changes for that
library. In the meantime, relax the test assertions and retry
snapshotting if we have a report node but are missing a property.

Multiply: starnix_coredump_diagnostics_test
Original-Bug: 131002
Original-Bug: 131004
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888437
Original-Revision: d1cdff226ec193ead787fe1e2e1b33fa1674bb57
GitOrigin-RevId: d385eacb65733e250e055bdec10dcd5c566e4732
Change-Id: I7ae3227687ac7a4302b03558f44bb54bffc8e732