[roll] Roll llvm-project/libc [libc] Remove test RPC opcodes from the exported header
This patch does the noisy work of removing the test opcodes from the
exported interface to an interface that is only visible in `libc`. The
benefit of this is that we both test the exported RPC registration more
directly, and we do not need to give this interface to users.

I have decided to export any opcode that is not a "core" libc feature as
having its MSB set in the opcode. We can think of these as non-libc
"extensions".

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D154848

GitOrigin-RevId: 6e4065e9bf4e890ae3f539e87952168725e47132
Original-Revision: 522978c2a865db7660d27a30554b5acd19d305b7
Change-Id: Ieb78de9c2f91c02a624fb394adb7b22a28687858