[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Remove test RPC opcodes from the exported header
This patch does the noisy work of removing the test opcodes from the
exported interface to an interface that is only visible in `libc`. The
benefit of this is that we both test the exported RPC registration more
directly, and we do not need to give this interface to users.

I have decided to export any opcode that is not a "core" libc feature as
having its MSB set in the opcode. We can think of these as non-libc
"extensions".

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D154848

GitOrigin-RevId: 4fea03968b597d0fbe5575e969b56452eef66213
Original-Revision: 522978c2a865db7660d27a30554b5acd19d305b7
Roller-URL: https://ci.chromium.org/b/8774895014341209489
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888572
Original-Revision: 739b0954623ec0419572c92b620bc53ba112e919
Change-Id: Ie789a24cf2b7af1dba24a9a414dfc7691ebaefc1