[roll] Roll fuchsia [tee_manager] Add core shard target
Adds a target for the "base" core shard.

Before, the base core shard would be included in a board-specific
shard and extended with `use` declarations to route protocols to
v1 components or the shell realm. The derived shards only have `use`
declarations, and nothing else.

Once https://fxrev.dev/881091 is submitted, the `use` declarations
have no effect and can be removed. That makes the derived shards
equal to the base shard, so a follow up change will replace them
with a reference to this new target.

Once the derived shards are removed, the `base` core shard can
be renamed.

Original-Bug: 104628
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/885774
Original-Revision: 333359494506927fc4afed317d959fc6d395710c
GitOrigin-RevId: b80adcc6444d329c4e844b2cd7e90939fc858f2e
Change-Id: I5b2455ff0998c9e2afcf4e9c1a0209f48403e5ab