[roll] Roll fuchsia [netlink] Support RTM_NEWROUTE e2e
This change allows netlink sockets/clients to perform the `RTM_NEWROUTE`
request.

For now, we assume that if the route in the request is valid, we can
attempt to add the route to the Netstack. In other CLs, we will
early-exit route requests if it is not unique from the routes already in
the routing table.

Support for using the `NLM_F_REPLACE` flag will come in an additional Cl
and implemented if/when it is needed. b/290803327

Original-Bug: b/283136222
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878533
Original-Revision: ff72e58de481ac2264678fa7e296595fa5cfa260
GitOrigin-RevId: 39ae76f65ad5ec3eaf8104a80703bfc10eaec347
Change-Id: I955715bc7c5f9408a0c6fc38d028c9816ad58386