[roll] Roll fuchsia [debugger] Do not assume gTest during `run-test`
This flag is gTest specific, and causes other types of tests like gUnit
tests to not execute correctly in the debugger.

Replace the corresponding `run-test` case with a script. We no longer
expect to hit implicit breakpoints from a test case failure in the
inferior, but this does ensure that we can start a test and intercept
the output.

Original-Bug: 130870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887792
Original-Revision: e4aa90b95340b56aedea37acb8bfafb4ea449352
GitOrigin-RevId: 3e10ed9308e9af5f797bd7d9355fdebb8df4215c
Change-Id: I293784b0282e4f66b3f7c95b6baebec60c5dc46c