[roll] Roll fuchsia [wlan][hw-sim] Implement realm factory configuration
Extends the realm factory FIDL protocol to allow the test suite to
configure the test realm.

Note that this was tested on a future commit together with the hw-sim
tests that actually require configuration, and the test in
`realm_factory/test.rs` was deleted.

Original-Bug: 125740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882913
Original-Revision: daabd492c86db1f6ae0c578c7cb263492fb35c52
GitOrigin-RevId: 474c5a0ec13348dd37a90b689d5c3f6556f96bae
Change-Id: I7e9206060445cc9c8edfcefc37bf3070a98a3aef