[roll] Roll fuchsia [profiler] Restructure Target Management
The method of adding targets during a run was... bad to put it lightly.
This change the approach to something much safer. It forces a requery
into the tree of targets each time instead of trying to save a reference
to the target a watcher is watching.

Since this adding is always done on a single threaded async dispatcher,
we don't need a lock protecting it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884348
Original-Revision: 3e1120b6b4019d927090415f764b23c1876da750
GitOrigin-RevId: c877b6fea0c527c59ed902c977b7b2dd40b0a6f2
Change-Id: I6c55e1f4228197c0f50120844f4261a3ac38a1b8