[roll] Roll fuchsia [kernel][arm64][mmu] Add some ASID asserts
When context switching, assert that the asid_ of the ArmArchVmAspace we
are switching to is in the valid range for user ASIDs. Also add some
static asserts that user ASIDs do not overlap with the UNUSED and GLOBAL
sentinels.

When the ArmArchVmAspace is destroyed, the asid_ is set to
MMU_ARM64_UNUSED_ASID (0), so this change prevents accidentally sharing
TLB entries across processes. This should not happen anyway because no
threads should be running after the ArchVmAspace of the process has been
destroyed, but the additional safety check does not hurt.

Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884754
Original-Revision: b3e06907c06ca3601cb81bc8f53567d559488dcf
GitOrigin-RevId: 18a2125b354e77c3774c0919733b4f83a2db9ed3
Change-Id: I30824b5304a915f7616fd12d39f470dedacef57a