[roll] Roll fuchsia [fxblob] Create core_with_fxblob product
We're transitioning from fxblob being a board config to a product
config, now that technical roadblocks in assembly have been removed
allowing images configs to come from the product.

For now, this config is more or less a stub and won't work on its own.
It needs to be used in conjunction with the x64-fxfs board.  Eventually
it will be self-contained.  It's being defined now so we can do a soft
transition of existing CQ builders.

Note that the x64-fxfs board is unchanged by this patch and works with
or without the new product.

Original-Bug: 291976815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890177
Original-Revision: 4a4f1df3f9e574afc88fd0421c3b443a1a5d0525
GitOrigin-RevId: a628142cc74122cc0e6c4d071608010015ca9b8a
Change-Id: Ib3b879b81ffc463bae873ef5a4774a668dcb6541