[roll] Roll fuchsia [netstack3] FakeNetwork holds a weak device ID
In some tests we create a FakeNetwork that holds strong references to
ethernet devices created in the test. This can make it tricky to
properly clean up the devices at the end of the test and avoid dangling
strong references. Change these tests so that FakeNetwork instead holds
a weak device ID that it upgrades as needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889874
Original-Revision: b59eb741300cebf528f220309c5d2f6b2af7f865
GitOrigin-RevId: add33e59528ee9ddcd206088ad12d530ba47b47b
Change-Id: If7a4274348541c6b7b28515f7a7dd4b86a0cd56b