[roll] Roll fuchsia [storage] More Controller multiplexing fixes
This is another set of generic fixes to either use the correct
FIDL channel or connect to the fuchsia.device/Controller path.

This causes a handful of storage tests to pass with multiplexing
disabled, we are down to only a couple of failing tests now.

Original-Bug: 127870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889879
Original-Revision: 17bbb926805db4e1bb38e8274c24ce54d24873de
GitOrigin-RevId: 29d0e094abbf21b16fca740fd6ff9ad26e054c98
Change-Id: Ic506e0dfe59ec37c94ca80b79607908769565f31