[roll] Roll fuchsia [starnix] Cleanup waitpid tests global state.
This CL modifies the WaitpidExitSignalTests so that the global state
they use is cleared before each test is run.

These tests spawn child processes/threads and wait on them, checking
that they issue the right signals when they exit.

The previous code had a bug in which it was possible for a signal to be
received before the fork was created, causing the global state to
contain bogus signal information.

To mitigate that, we now clean up the global state after we fork, making
it impossible (modulo starnix bugs) to receive unwanted signals.

BUG=131156

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890116
Original-Revision: 1072ad049c5375cb0978d16280a69e266bacd45f
GitOrigin-RevId: f6772f2ae090b773d06404677d18bee28b2a8202
Change-Id: If0e3601f36e9dedb78c790aed82a6ca2810f0820