[roll] Roll scudo [scudo] Make Options a reference for functions.
Modify all places that use the Options structure to be a const
reference. The underlying structure is a u32 so making it a
reference doesn't really do anything. However, if the structure
changes in the future it already works and avoids future coders
wondering why a structure is being passed by value. This also
makes it clear that the Options should not be modified in those functions.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D156372

GitOrigin-RevId: b157c3ab72273fbe76fc40ea80f8968c40716811
Original-Revision: 597231560e99e00c115239bde5613889a035346f
Change-Id: Ibb29200caaf7281244b9d9045643fd48e820a2b3