[roll] Roll fuchsia [superproject] Roll scudo [scudo] Make Options a reference for functions.
Modify all places that use the Options structure to be a const
reference. The underlying structure is a u32 so making it a
reference doesn't really do anything. However, if the structure
changes in the future it already works and avoids future coders
wondering why a structure is being passed by value. This also
makes it clear that the Options should not be modified in those functions.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D156372

GitOrigin-RevId: 065825f26f9d482f2c3bd7835798036d71e2cc5a
Original-Revision: 597231560e99e00c115239bde5613889a035346f
Roller-URL: https://ci.chromium.org/b/8774451345183747329
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891014
Original-Revision: 16af76afaee76d61c1278da2dd99335476cdf7b7
Change-Id: Ia0a470e438921980ba61336d13bfb3cfabd78835