[roll] Roll fuchsia [fuchsia-sync] Implement RwLock
We use two futexes, one for readers and one for writers. Once a writer
arrives, we mark a bit in our state so that readers do not starve
writers.

This RwLock does not track thread ownership for priority inheritance.
It's difficult to do that tracking in general for read/write locks, but
a future CL could improve this situation for states in which a single
thread has exclusive (writer) access to the lock.

Additionally, zx_futex_wake does not indicate whether we succeeded in
waking up a thread, which means we always need to broadcast the unlock
signal to all the readers whenever there are both readers and writers
waiting for the lock. If a writer wins the race, we will have woken up a
lot of readers without much benefit.

This lock will be used in Starnix. There is an argument that we should
not use RwLocks in Starnix. We will study our use of RwLocks in Starnix
and likely switch to mutexes in many cases. However, there are some
cases where a RwLock is needed by our current implementation approach.

Original-Bug: 88603
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/889013
Original-Revision: ab0aa185148fcccae149b1ded58c704f38b9677b
GitOrigin-RevId: 8823fa302ef70beba391602d5125c4ca7af9ce76
Change-Id: I4e3720d2127a00f73e08b615aa5be0ef525d9d69