[roll] Roll fuchsia [netstack3] Remove FakeBufferIpSocketCtx type
This is redundant with the FakeIpSocketCtx type, which it was just a
thin wrapper around. Replace all usages of FakeBufferIpSocketCtx with
FakeIpSocketCtx, implement any required traits on the latter, and remove
the now-unused type.

As part of this cleanup, remove the Default impls for some
UDP-specialized test types and replace usages with `new_xyz()`
functions. In addition to not causing trait impl conflicts, this is more
correct since the functions create types with devices and addresses,
which does not fulfill not the "empty" contract that Default implies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891174
Original-Revision: dff925a0a14747311446b244794e5a3e7691669b
GitOrigin-RevId: fcb6599344a6bbdb01b39020c587e8ddf8f0f547
Change-Id: I569916027d5b70bfb737a154541cd0a6550a94e6