[roll] Roll fuchsia [netlink] Properly sync route test requests
Follow up to fxr/884054. That CL correctly blocked while waiting for
messages, but the messages may not have been sent at all due to
breaking from the `select` macro before 1) the route_set_fut could
send the message to the Routes Watcher and 2) the EventLoop could
process the event from the Routes Watcher, sending the multicast
message.

In the Netlink Route tests, combine the future that sends the
`AddRoute` message with the future that waits for X messages to be
received. Ensuring that both of these futures complete prior to
breaking out of the `select` macro will mean that the RouteSet handler
is able to send the event to the Watcher and the EventLoop is able to
observe the event from the Watcher.

Confirmed with @csuter and this seems to fix the issue with the
unification of the single and multi-threaded executor.

Multiply: netlink-lib-test-no-err-logs
Original-Bug: 130548
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890669
Original-Revision: a7510455a6f04a6a9afe436ed5015f269b355a0e
GitOrigin-RevId: cf8b10d2b1ca10c27e248ce2c1335fc8a10c18e8
Change-Id: I3169acc654f66ef068ccfcca1ca42956b63f1583