[roll] Roll fuchsia [ffx][command] Move overnet init to proxy lib
This is a minor code re-organization to move the overnet initialization
code to a less awkward place. I'm not sure how it ended up here, but it
puts a heavy dependency chain on the ffx_command lib, which should be
more lightweight.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890647
Original-Revision: b9ab01f9688505471e09f6e706e2dbedfe2a1788
GitOrigin-RevId: e84d0a9e87b3b7170387dd853f52e7f5c9594ba5
Change-Id: Ia3d1a26b100298d0a057f7de5d3aa8e69b312cb6