[roll] Roll fuchsia [superproject] Roll cobalt-registry [networking] Modify socket option report params
Set the floor for the socket option report to 1 so that 0 is included
in the underflow bucket and we can tell which socket options are not
used at all.

Set the multiplier to 4 to increase the start of the overflow bucket
from 512 to 65536. This provides more accurate ranges at the top end.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/888993
Privacy-Approval: Cameron Dale <camrdale@google.com>
Original-Revision: f5abf6ed6607ec218f8dd6862281737bab869389
GitOrigin-RevId: 71be0f46861963e6c53ec55fcb346c14542f0e97
Roller-URL: https://ci.chromium.org/b/8774372493927835777
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891578
Original-Revision: 2c660e1336f97543064e298c75748a338271a8f3
Change-Id: I2a93254d838fd05cb9452eb2ccb3d52cabaeb8f5