[roll] Roll fuchsia [fxtest][rewrite] Terminal output library.
The termout library provides support for repeatedly overwriting the
contents of a terminal window to create animated status updates, without
scrolling the window.

We use this instead of curses because curses switches to an entirely new
buffer, and we want to preserve the context of the user's terminal
history.

This library uses ANSI escape codes to move the cursor, delete
appropriate lines, and overwrite them to create the desired effect.

It includes a termsim module that simulates a small subset of ANSI
escape codes for testing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884706
Original-Revision: 2f37f07be705ad81e0e0d17e78c75316caf13a9e
GitOrigin-RevId: 92c5ff2f3b1cfc74ba527d30bc75681720abd97c
Change-Id: I009ccdbe33d23e90cd2cacacbe96172e10c93e4d