[roll] Roll third_party/go [fidl] Don't close channel on 1way messaging errors
When sending a one-way/fire-and-forget message, do not close the channel
if an error occurs. This is so that a client can read messages sitting
in the channel after the error occurs. It should also be noted that for
one-way/fire-and-forget messaging, there is no guarantee of message
delivery or handling.

This change is needed for Ib72a8a7610637731a518241d9d048be9bd4fdf1a.

Original-Bug: b/292622684
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/891252
Original-Revision: 2b3f72fa1f4168360d586ca7a8428256dc0acfa6
GitOrigin-RevId: 3e9bb1531e236e1e7968969ae8c8eb35ac0ef50d
Change-Id: I72ece0d7b3896f1039e5fb82b64f0ba195f8cf6e