[roll] Roll fuchsia [vim3][dwmac] Fix interrupt behavior
If an ethernet cable is plugged in on boot, the DMA_STATUS_GLPII bit may
be set. That will cause the interrupt line to always be high, so with an
edge-triggered interrupt the interrupt wait may take forever.

There are two parts to this fix:
* Mark the interrupt as level-triggered. That matches the documented
  behavior of sbd_intr_o.
* Read DW_MAC_MAC_LPI_CONTROL as needed to clear DMA_STATUS_GLPII.

Original-Fixed: 131163
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891177
Original-Revision: fe3ebf851a4cdb7a704126d05e2ed19fbbf83e1a
GitOrigin-RevId: 471ba8ef50977fcf3bde2c5e5687531d6fa5ee2b
Change-Id: I25c9ed06af383fac2925efacabd487f7a3965128