[roll] Roll fuchsia [netstack3] Rename SocketStateEntry::try_insert
Rename try_insert to try_insert_with since it takes a callback that is
used to construct the inserted value. Use the name for a simplified
version that provides a static id. Note that try_insert_with is not long
for this world, but add tests for its behavior on error (which seems to
have been previously untested).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891205
Original-Revision: f7cf6d3b783e26d94657011abfb0d4b04e3812c5
GitOrigin-RevId: c7fe2c8df256da27b6d052d9810907d0a2032b17
Change-Id: I3e2259fd4023bcc380224ff51e4e700882670531