[roll] Roll fuchsia [netlink] Fix panic opportunity
`netlink` assumed that `AddressStateProvider.WatchAssignmentState` will
always be handled before the address is removed for reasons not related
to add-errors. This is not true because in theory, the request could be
sitting in the channel when the address is removed.

This change introduces an "on-added" event to the ASP that the server
will emit when the address is successfully added.

Original-Fixed: b/292622684
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890667
Original-Revision: 70440d65e5423e36871672bc45a1929c14aa6d7f
GitOrigin-RevId: 66d0e965a35b1a1c5d5999094556e9c306fce4c1
Change-Id: I237435b68551ed5c37d13d39abed257f31d1f3e6