[roll] Roll fuchsia [kernel][arch] Support values of TCR_EL1.TnSz != 24
On ARM, TCR_EL1.T{0,1}Sz are configured as
`64 - the size of the virtual address range`. These values must be
accounted for in the paging traits, as beyond restricting the size of
the virtual address space, they may further restrict the number of
paging levels (all else held equal) - and the number of paging levels is
a static determination made by the traits.

The relevance of this support is that we currently support lower ARM
virtual address spaces of size 39 and 31 (both giving three-level
paging) in physboot and the VMM.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887018
Original-Revision: a33d85fa273e86d1cbda92e273db337724dc8075
GitOrigin-RevId: c992cd46d667826f2c75c900254cb7452f816e78
Change-Id: Ie9725b85b25ad47fe1a856f4b01a15d4cef4801b