[roll] Roll fuchsia [netstack3] Provide summary for socketmap addr tag
Instead of making the entire address vector available to socket
implementations for the purpose of computing the per-address tag values,
only provide a summary of the information. The tag computation doesn't
need to know, for example, whether a connected socket with a device
bound is bound to a loopback vs ethernet device, just that there is one.

This has the happy follow-on effect of letting us get rid of the
`Bound<S> : Tagged<AddrVec<I, D, A>>` bounds that were annoyingly viral
and therefore everywhere. Since there is a blanket impl in terms of `S:
SocketMapStateSpec`, that bound can now just be inferred.

Original-Bug: 21198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891172
Original-Revision: d568c462817915921372043967224aa39ed9b08a
GitOrigin-RevId: f47bf3204de7e9416d89c03f85df53946a5fbc6b
Change-Id: I3f82f773f61db687ca4b3bf927421647341149be