[roll] Roll fuchsia [netstack3] Hold v4 & v6 socket in UDP v4 map
Modify the datagram spec impl for UDP to vary the type of receiving
identifier by the IP version of the demultiplexing map. Allow holding
only a SocketId<Ipv6> in the IPv6 demultiplexing map, but an
Either<SocketId<Ipv4>, SocketId<Ipv6>> (spelled as EitherIpSocket<Udp>)
in the IPv4 demultiplexing map.

This is a key first step towards enabling cross-IP-version reception of
traffic since it will enable IPv6 receivers to be found for IPv4
packets. While this change updates the types to make it possible, there
is no path for adding IPv6 sockets to the IPv4 map so all socket IDs are
still assumed to be IPv4.

Original-Bug: 21198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891173
Original-Revision: 8c1acab3a1ec0b677fd6bbc0e62b587c0e2c313c
GitOrigin-RevId: 7a63e781721d19bb3c82c476d4df116172dda439
Change-Id: I690135384cd97a4ad2cc9996888b2d2869067322