[roll] Roll fuchsia [profiler] Migrate ffx profiler to a subtool
ffx profiler depends entirely on fuchsia.cpu.profiler which is
currently an internal API, so it shouldn't be distributed as part of
the ffx sdk.

This migrates it to solely be an ffx tool, which will also benefit its
iteration time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890856
Original-Revision: fbdac000abb2d4a4de5ad4b63ff40d39547ff6e8
GitOrigin-RevId: a5095a5e286947b9b6cddfa1470ed2662e7f41a2
Change-Id: I8e65560563053461e897f8eecf35904196935a65