[roll] Roll scudo Revert "[scudo] seperate cache retrieval logic"
RoundedSize is supposed to be used on directly mapping. To determine if
a cached block is feasible, the size doesn't need to be rounded in
advance. As a result, the use of RoundedSize may miss some chance of
using cached blocks.

This reverts commit 4c6b8bb87b3452d0bcef83cd0ea712d8426603b8.

Reviewed By: frs513

Differential Revision: https://reviews.llvm.org/D156583

GitOrigin-RevId: 86973bd921de077d783a1d725ab676cee4d8bae1
Original-Revision: d7a6bb32e0bc6f3204ccb5f1cea87f34002ab807
Change-Id: I52a0292e824d0fd23b67cd3f59a8b33167f06e4b