[roll] Roll fuchsia [check-licenses] Skip ".cache" directories by default.
This is causing issues in some environments.

For the most part, check-licenses shouldn't be traversing any directory
that starts with a dot, but the skip logic doesn't support wildcards at
the moment, so this CL just adds ".cache" to the list of skips.

Original-Fixed: 130958
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892659
Original-Revision: 83e882af642ffcd54acc63c433ee9b09c2bb5345
GitOrigin-RevId: 0a282dc391aa9f8bf0c209c1c5d1ec781e3aa036
Change-Id: Icc28ea4972c6dfb3c787a630969ee0f85e362f2e