[roll] Roll fuchsia [symbolizer] Support backtrace prettifying
Add a new flag to the symbolizer, --prettify-backtrace. When enabled,
backtraces in {{{reset:begin}}} and {{{reset:end}}} pairs will be output
in a compressed and prettified format, allowing more readability.

The flag is not enabled by default and there should be no behavior
change when the flag is not enabled.

Original-Bug: 59736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892726
Original-Revision: 035a10194e1a4c75d3fd7396a04b17c70f18ea43
GitOrigin-RevId: 356a14351c0b57ec51ee731096a112ed9199f7e0
Change-Id: I667f5fc725ca652c027f497d4bb65143001953e7