[roll] Roll fuchsia [overnet] Delete images for overnet serial tests
As mentioned in fxbug.dev/80059 overnet is no longer used, and if it
comes back it will look completely different. Delete the images used
in the overnet serial tests so that they do not need to be maintained
or else bit-rot.

Original-Bug: 80059
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892878
Original-Revision: 4ecc6095ab63fd64021c130ddd3c5f203bdd78b0
GitOrigin-RevId: 4ef2697aa61794dd3f4fc8197f89f03d3a2281f2
Change-Id: I7cd09387c2ef0768df2b11ed30d81cb63d3e378b