[roll] Roll fuchsia [starnix][tests] Fix flake in test_clock_nanosleep_interrupted_relative_to_fast_utc_clock
This test was attempting to interrupt a relative sleep from another
thread roughly halfway through the sleep. Since the test can't see the
time value used by the relative sleep it is not possible to reliably
know when the interruption will occur. This updates the test to avoid
doing two relative sleeps and relaxes the invariant so that if the
interruption happens earlier than expected the test still passes so long
as all other invariants are maintained. This also improves the assertion
output on failure to aid future debugging.

Original-Fixed: 131337
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892736
Original-Revision: 714b77a0754ae94ce32513b8edd4b1fead1d7529
GitOrigin-RevId: 714cab76abcce5eaf3f822befce25f01f029359c
Change-Id: I84baf8ee78c38cd05b0e8a2639f19a9b2c205d2f