[roll] Roll fuchsia [ld] Load main executable
This makes the dynamic linker actually handle loading the
executable.  This includes setting some of the groundwork for
loading modules in general, but does not yet do any relocation.
After loading, it hands off to the executable with the same
initial conditions it started with.

In the POSIX incarnation, "loading" is actually just working with
the executable image already loaded by the system program loader
and found via auxv entries.  The test harness is enhanced to
mimic the system program loader more fully by loading the
executable and filling out the related auxv entries.

The build and test machinery is prepared for a wider set of test
modules to be built in the future.  This is a step closer to
generalizing the same tests to be done in the current in-process
fashion as well as in multiple out-of-process fashions.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888711
Original-Revision: 23d2b8d98365bf95087dc1fda9d0ce4563c7b860
GitOrigin-RevId: 638b1b749ca01907a56504bcbdbc1ee4c772c856
Change-Id: I22535443360666d4b0d23b5e33d1bb1061b2e8b1