[roll] Roll infra/recipes [sdk] Move SDK upload back to after companion images and not before
In an effort to modularize the core and Bazel SDK steps, I'd mistakenly
moved the uploads ahead of the companion image uploads, which breaks the
expectation of downstream users that the companion images are ready by
the time the core and/or Bazel SDK is uploaded and tagged in CIPD.

This change restores the uploads back to their ~original phase in
execution.

This change doesn't affect the GN SDK upload because it is effectively
unused aside from a script in fuchsia.googlesource.com/samples that does
not rely on companion images (see https://fxbug.dev/130260).

Original-Bug: 49542
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/894138
Original-Revision: 701a1cbea57e616d73256fb72209f3b21e733860
GitOrigin-RevId: 04d8e15f31faaeb2f4c155a0896cec3996a3f3de
Change-Id: I0712ea98f98ee1402a8bcac315c161610a101002