[roll] Roll fuchsia [netstack] Warn for not-found thread profile
Currently we log an error whenever we fail to set the thread profile
for the threads the netstack runs on, but some failures are expected
when no matching thread profile is configured with the requested
profile name. Downgrades that specific not-found case to a warning
instead of an error to mitigate logspam.

Original-Bug: b/294198743
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893960
Original-Revision: 67a46b185aeecc2f3b43b2bb2da243fdb63ce00e
GitOrigin-RevId: 6605330d05ef2b54c2f0dc85fddd3fba2dbce566
Change-Id: I2cae13476d7d83637bcff0f3a16b7f201945502f