[roll] Roll fuchsia [starnix][pidfd] Integrate with waitid
Previously, pidfd was holding onto the Task with a WeakRef<Task>, but
that doesn't work for waitid, which uses the pidfd to specify the task
after the task has exited. After this CL, we use the pid itself, which
is sufficient for now but will need to be revisited in the future. See
the comment on `PidFdFileObject::pid` for more details.

Test: pidfd_open04
Original-Bug: 119476
Original-Fixed: b/287120994
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893766
Original-Revision: 65866f583030ef26a1968eba2f3bbc71045286b3
GitOrigin-RevId: 1ff23c7bb69740b36cca0e7f50d8c3eded685764
Change-Id: I00de96203234e1653d48f4814b1ee02ab9838f64