[roll] Roll fuchsia [fxfs] make XtsCipherSet on demand
Instead of pre-making the XtsCipherSet, make a new one every time we
need the keys.

Premaking the set means we create the Xts128 struct containing two of
the Aes256 structs, which each take up about 1k of memory regardless of
if we have AES intrinsics on the board we are running on. This is an
enormous amount of memory per node, so instead we pay a cpu cost every
time we encrypt. The cost is very small when intrinsics are enabled
(just a copy or two), thought it is slightly larger when they aren't.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893460
Original-Revision: 3a2287ff43875877d3b1a24ca2eda6a49623aa3a
GitOrigin-RevId: 43bf59640acfe4932fb0836a25ac8942079b3748
Change-Id: Ie03f9530fe7d8cd87771f29f8990fa3ce8c3ea61