[roll] Roll fuchsia [intel-i915] Fix AcpiMemoryRegionTest setup.
AcpiMemoryRegion supports two construction modes:
* The "production" path takes in a VMO and a buffer mapped from the VMO
* The "test-friendly" path takes in an unowned buffer

AcpiMemoryRegionTest uses the production path, but passes in a buffer
that was not obtained by mapping the passed-in VMO. This CL fixes the
test, and adds a debug assertion to the AcpiMemoryRegion to catch
similar mistakes.

The investigation and debug assertion are courtesy of fdurso@google.com.

Original-Bug: 131589
Test: fx test intel-i915-test.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894279
Original-Revision: 6608d38824d35b859ceac44b271011cc9b4d419f
GitOrigin-RevId: db4753cd89fe549e8c1a16849f78fdb8fe981c53
Change-Id: I02a86c4aabd38c89c5a4fc2b2181182890a16db6