[roll] Roll fuchsia [amlogic-display] Fix a naming confusion in canvas endianness.
Canvases are default to little endian for each 128-bit chunks,
and captures are configured to big endian for each 64-bit chunks.
In order to read the captured bytes correctly, we need to set the
canvas endianness to match the capture interface.

This fixes an incorrect naming of canvas endianness used for captures.
The capture should be 64-bit big-endian instead of 64-bit
little-endian.
This change also replaced the magic numbers with bitfield names.

This change doesn't change the behavior of the driver.
Test: no-op.
Original-Bug: 131232

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894290
Original-Revision: ee252855331f1a8fe8acedeb7a7d566291c2b768
GitOrigin-RevId: 67fe44df47097217106b0fae4624e2569410534d
Change-Id: Ie49d1a1b5eb60fb1e95eed94733a09236053b6d0