[roll] Roll fuchsia [py] Use argparse.BooleanOptionalAction to remove --foo/--no-foo
This commit updates the argparse parser to use the BooleanOptionalAction
action, which was introduced in Python 3.9. It was possible due to the
Python upgrade we made to 3.11. This allows users to specify the foo
flag with a single flag, rather than having to specify both --foo and
--no-foo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894388
Original-Revision: 0300a532e17421ae428c22b2796dd4a5c122b6a5
GitOrigin-RevId: 32076f5eb8e62c4e3ff09b4834d7290253068edf
Change-Id: I910a9cc33f4abe0db8cf10833e001fa5abd40864