[roll] Roll fuchsia [fx][serve] Don't fail if server address is empty
This changes `ffx-repository-server-running-port` to return an empty
string if the server is not running. This matches the behavior of
`ffx-repository-server-running-address`.

In addition, this removes an unneeded feature flag, and puts quotes
around variable expansions to distinguish between empty values, and
values with whitespace.

Original-Fixed: 131459
Original-Bug: 120327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893382
Original-Revision: 771b7df6ee2007f56748778c61d0af2ac9a3d770
GitOrigin-RevId: 105a8bff46638a1a5a640d6a882f4aaf820b0553
Change-Id: I3b58beed330223e315eb6a156c0c3a5513a2662f