[roll] Roll llvm-project/libc [libc] Add support to compile some syscalls on 32 bit platform
This patch adds a bunch of ifdefs to handle the 32 bit versions of
some syscalls, which often only append a 64 to the name of the syscall
(with exception of SYS_lseek -> SYS_llseek and SYS_futex ->
SYS_futex_time64)

This patch also tries to handle cases where wait4 is not available
(as in riscv32): to implement wait, wait4 and waitpid when wait4 is
not available, we check for alternative wait calls and ultimately rely
on waitid to implement them all.

In riscv32, only waitid is available, so we need it to support this
platform.

Reviewed By: michaelrj

Differential Revision: https://reviews.llvm.org/D148371

GitOrigin-RevId: 0950f57c92b7d358abd96f47d14a28bcab35a1de
Original-Revision: e21d877d45bcbce28339ba2107709130e5594f46
Change-Id: I6fbea0c46a7ff421f353afdfe951981ebbb3fb31