[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Add support to compile some syscalls on 32 bit platform
This patch adds a bunch of ifdefs to handle the 32 bit versions of
some syscalls, which often only append a 64 to the name of the syscall
(with exception of SYS_lseek -> SYS_llseek and SYS_futex ->
SYS_futex_time64)

This patch also tries to handle cases where wait4 is not available
(as in riscv32): to implement wait, wait4 and waitpid when wait4 is
not available, we check for alternative wait calls and ultimately rely
on waitid to implement them all.

In riscv32, only waitid is available, so we need it to support this
platform.

Reviewed By: michaelrj

Differential Revision: https://reviews.llvm.org/D148371

GitOrigin-RevId: d3b73535bc41f1f96dbd9b2382f01abc3a55e8dd
Original-Revision: e21d877d45bcbce28339ba2107709130e5594f46
Roller-URL: https://ci.chromium.org/b/8773770870952655153
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894858
Original-Revision: f0917d9d0800d247fe974b6685af1eb88edaf9f2
Change-Id: I6668c5cd6dbbdce432faea8791c8ff5d86139ef3