[roll] Roll fuchsia [settings] Use fasync::Time to avoid test flakes
DeviceStorage::with_stash_proxy was using std::time::Instant::now()
and then converting this to a fuchsia_async::Time when creating the
Timer for the next flush. The conversion occurs within the WakeupTime
impl for std::time::Instant [0], which involves another call to
Instant::now(). If the system monotonic clock advances in between the
creation of the initial Instant and the conversion to the Timer, then
the Timer will be scheduled for less than MIN_FLUSH_INTERVAL_MS in
the future relative to the executors's time when the Timer is
registered.

This was causing test_multiple_write_debounce to flake on asan builds,
because sometimes more than a millisecond would elapse between the
creation of the two Instances, and the test advances the executor's
clock by MIN_FLUSH_INTERVAL_MS - 1 and expects the flush Timer to not
trigger.

With this change, fuchsia_async::Time::now() is used directly instead
of first creating a std::time::Instant. fasync::time::now() gets the
current time from the executor, so during tests it is guaranteed that
the flush timer will always trigger exactly MIN_FLUSH_INTERVAL_MS
(as measured by the executor clock) after the preceding flush.

[0] https://cs.opensource.google/fuchsia/fuchsia/+/main:src/lib/fuchsia-async/src/runtime/fuchsia/timer.rs;l=31;drc=babf24bb25a3d8e768b25b3133ee5cb8e4f69b3e

Original-Bug: 131240
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894010
Original-Revision: 822f954a507e9125d23727f98f81f566fecf5a28
GitOrigin-RevId: 505b19682dab36b3b22e98e904ef901fc96e55e9
Change-Id: Ieaab0b29b50e915f9aacceeba010fd9604e6868c