[roll] Roll fuchsia [netlink] Support RTM_DELROUTE e2e
This change allows netlink sockets/clients to perform the `RTM_DELROUTE`
request.

Re-arranged some error handling, allowing for the unicast field
extraction function to be used for NEW and DEL route requests. This
required updating some Err values in existing tests due to hitting the
Err flow for EINVAL with an inadequate destination subnet rather than a
missing Oif NLA. This new implementation is more appropriate because if
an Oif is not present, we will need the subnet to be appropriate so we
can resolve the destination address (see b/292103361). Doing this
resolution is a concern for NEW requests, but not DEL requests.

Extracted implementation and helper code into generic over IP functions.

Original-Bug: b/283136222
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/890656
Original-Revision: fa5515e92ee74b435b20710dc1d3c729aa35e950
GitOrigin-RevId: 3740405c8ef7ca2912409e57d1e98dc246311505
Change-Id: I5ce71b2da7cd178c8afa41becfd367a5321105a3