[roll] Roll fuchsia [starnix] Switch back to parking_lot::Mutex
This CL switches Starnix back to using parking_lot::Mutex so that we can
study the performance effects of using fuchsia_sync::Mutex. The CL also
makes both the Mutex and RwLock backend configurable with a GN variable.

We still use fuchsia_sync::Mutex when debug assertions are on because
that mode is not performance-sensitive.

Original-Bug: 88603
Original-Fixed: 131092
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894801
Original-Revision: 052bd302c5f3a80e68e9b31e5253622515810368
GitOrigin-RevId: 27439c8f06a50d4e80c146e590013410f0944b88
Change-Id: I2c958a7c224b73697bd6dc526d87d35e6cfaa864