[roll] Roll fuchsia [libc] Move abort into llvm-libc implementation structure
This moves the abort implementation from the old musl fork code
structure into the new llvm-libc-based code structure.  The
implementation is unchanged except for the never-raelly-reached
fallback case using the ZX_TASK_RETCODE_EXCEPTION_KILL value
instead of -1 in the zx_process_exit syscall.

What's really new here is setting up the libc build machinery for
having Fuchsia-specific implementations in the Fuchsia source tree
but structured as if they were in the llvm-libc source tree.
These implementations, of which abort is the first example, will
use the llvm-libc C++ namespace and boilerplate macro structure
for defining the code.  The llvm-libc test code from upstream is
used to test these implementations the same as the upstream
implementations for other operating systems are tested.

This resolves the failure of ld-tests in the coverage build.  Its
issue was that this build doesn't compile away calls to abort that
the other builds do.  We were linking in the abort implementation
from libc before to satisfy any such references, but because we
were using the old style of libc implementation code that defined
the public API symbol with default visibility such that references
require symbolic resolution in dynamic linking, which should never
be required by references in the dynamic linker except those to
the vDSO (not even references to itself like this one).  Now the
abort implementation comes from the "hermetic" compilation of the
libc code, which uses hidden visibility instead so that no dynamic
linking is required.

Original-Bug: 130483
Original-Fixed: 131578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894361
Original-Revision: 08afa14c2bf0359009dfc51b6aec341f45d76d60
GitOrigin-RevId: d45ffcc300dfbc19bc70b4402f1f74ff1677a671
Change-Id: I1d35c45c3c9b9d65a822377efe8bb903ff6dc162