[roll] Roll fuchsia [diagonstics] Emit system-update-committer logs over serial
This includes the system-update-committer logs in the serial logs. This
is useful because system-update-committer is responsible for deciding if
we should accept or reject an OTA. If the OTA is rejected, it might
trigger the device to reboot before the netstack has come up. When that
happens, it's difficult to capture the system logs through `fx log` or
`ffx log`.

Original-Bug: http://b/294114858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895079
Original-Revision: 152496151c03186788fd5a835cd02a9d45f176d4
GitOrigin-RevId: a1a1a844b72cc5fa7eddeed0daf89bf1fcffff7a
Change-Id: Ibf1a554abe87fcea6d59d6dc99d78db91f5f59da