[roll] Roll fuchsia [starnix] Assert SO_SNDBUF/RCVBUF is doubled
Add a test which asserts that SO_SNDBUF and SO_RCVBUF is doubled on
get compared to the value that is set, which is the desired Linux
behavior.

Currently NS2 exhibits this behavior, but NS3 doesn't, so this test
acts as a change detector for when we switch over to NS3.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893378
Original-Revision: 01ad8417ccf670ea8870f5ab7233a2a9f80a2d1a
GitOrigin-RevId: ccc5901f9937acd1a94b49ddc25a93512e2a176e
Change-Id: Icb1ac30a4aa92228a7c408448ad0463cdc4542e2