[roll] Roll fuchsia [diagnostics] Efficient histogram encoding
Histograms have used an array of Bucket {floor, ceiling, count}. This
takes a lot of space and is very inefficient to deserialize because it
creates many objects.

Histograms are changing to use parameters + array of count. If more than
50% of the counts are 0, the array of count will be sparse, and
accompanied by an array of corresponding indexes.

This change affects the ArrayContent struct in Rust, the output of
iquery, the format of the ArchiveAccessor ABI, the contents of
inspect.json in snapshots, Sampler, the Inspect Validator, and a
change detector in sdk/ctf/tests/fidl.

Original-Fixed: 125876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875538
Original-Revision: 5f7a9d1a7f2ba9dba5e02b2bb96659b5f922f452
GitOrigin-RevId: dc0581f1583792abd8dbeefb0c32c8a120c30bfe
Change-Id: I422450fbfb7b00d6f5055a6de2f0aaf24a8398f2