[roll] Roll fuchsia [kernel][test] Convert an ASSERT to an OOPS.
When the kernel unittest framework runs a test suite, it used to
record which CPUs were online/active before the test suite started,
and again after it was finished.  Then, it would ASSERT (kernel-panic
assert, not test-assert) that sets had not changed (IOW - if a test
took a CPU offline, it turned it back on again before returning from
the test).

The trouble with this is that if a test fails and does not clean up
after itself, this ASSERT fires typically well before the DLOG drains,
and we loose the test results, making analysis tricky.  Did the test
fail, or think it succeeded?  If it failed, why did it think it
failed?

So, change this behavior.  Instead of performing this check per test
suite, instead perform it individually for every test in the suite
(helping to isolate which test may have messed things up).  In
addition, instead of ASSERTing, print a KERNEL OOPS if the
before/after online/active counts don't match, and _also_ force the
test status to failed (even if the test would have otherwise thought
it succeeded), making sure that things like CI/CQ are sure to pick up
on the failure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896005
Original-Revision: 5735805036a18c2eb8de0bac9d056319748085f1
GitOrigin-RevId: 99ce8dea720bb61a6896833c44d5f9011c021275
Change-Id: Idec9ab5b91cb5a7344f4c0ea526941fb915e2cb9