[roll] Roll fuchsia [ld] Refactor tests
This reconsolidates the Launch object into the test fixture, so
that there is a different test fixture with the same API for each
version of the tests and the fixture provides methods to get the
test executable loaded and dynamically linked and then run,
however that happens.  Right now, that's a single test fixture
LdStartupInProcessTests but still using the TYPED_TEST machinery.
Soon this will be expanded to include separate-process versions
of the same tests, each with a different test fixture class with
which the same TYPED_TEST cases can be instantiated.  Later that
will extend to having the same tests use out-of-process dynamic
linking instead of any kind of use of the startup dynamic linker.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895580
Original-Revision: fe9433e3f7bdd0210dc977d7cd7ac7d98556e630
GitOrigin-RevId: 927c40e42a2c0d0b867fe9c7577b8577112969de
Change-Id: Ib6007b0a59697289857bedc01264cc82afadcaec