[roll] Roll fuchsia [wifi][brcmfmac] Fix wifi metadata retrieval in DFv2
Give wifi metadata to wifi composite device instead of sdio composite
device. Forward wifi metadata from wifi device to `brcmfmac-wlanphy`
child device so that the driver that binds to `brcmfmac-wlanphy` has
access to the wifi metadata when trying to retrieve metadata. This is
all done so that the brcmfmac driver can retrieve wifi metadata
correctly in DFv2.

In DFv1, the driver for `brcmfmac-wlanphy` could get the wifi metadata
without needing this change because DFv1 gives access to all metadata of
all ancestors of the device the driver is bound to. In DFv2, a DFv1
driver only gets access to the metadata of the device it is bound to.

Add platform device fragment to wifi composite device so that its driver
can retrieve metadata. DFv2 does not support composite drivers getting
metadata from the composite device it is bound to.

Original-Bug: 131219
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891204
Original-Revision: b5df63fc9490f1a79b264a91ae4ab5e05b2b8465
GitOrigin-RevId: c4e153578b3a8e4fdcaf01bd3d3234ce9b96072e
Change-Id: Iefa6b886ba848193ef37bfafedef53bc8b55461f