[roll] Roll fuchsia [sys] Improve ABI compatibility errors.
Instead of talking about ABI revisions, which are difficult-to-parse
random numbers, the error message will try to talk about API levels
when it understands what API level the component was targeting.

When the ABI revision is unknown, it suggests that the problem is
likely that the OS is too old.

Original-Bug: 129346

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/893945
Original-Revision: ae951d02d2fe5fb5c47addb599ccaed31aab7067
GitOrigin-RevId: 421cf377589b8fbdf4298084020e3503505ece92
Change-Id: I46300eca6d43502927e6ff3be16f69fdcd50c6f9