[roll] Roll fuchsia [sdk][fuchsia.data] Increase MAX_VALUE_LENGTH from 8192 to 32768
In I36dd5c08eab30d0fc37820d9d5324e2328639882 we increased it from 1024
to 8192. Now we're again seeing test names that go over that limit so
this CL increases the bound again.

Arguably, there is no intrinsic limit to a command line argument (on
Linux it could be as high as 2 MiB). So we might want to consider
dropping the limit, but that would be a breaking change since it would
remove a constant member from the FIDL file.

Original-Fixed: 294916548
Original-Bug: 131551

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895941
Original-Revision: c7bf0273b825ae4cd2cfda0ecfb23c85e5770a92
GitOrigin-RevId: 3ff7b5a8e913771b1b4ef98cacf8664dfce4851b
Change-Id: I9795c7566fe7376dc20b1d2adf5e65a600a7f380