[roll] Roll fuchsia [ui][a11y-manager] use `optional` in FakeColorTransformHandler
Using raw `bool` and raw `ColorCorrectionMode` makes the
tests hard to understand, because it's not obvious whether
an EXPECT-ation would fail in the absence of a call to the
handler.

E.g., is EXPECT_FALSE passing because the handler was
called in a way that caused color inversion to be disabled?
Or is it passing because the ctor initialized the bool
to false?

As it turns out, it is _not_ the latter, because
1. The handler uses the default ctor
2. The default ctor default initializes the `bool`, and
3. Default initialization leaves a bool uninitialized

That makes it too easy to write a flaky test (since
reading uninitialized memory yields undefined behavior),
and/or trip up the UBSAN bot.

Instead of using raw types, wrap these fields in
`optional`, so that the values are always initialized,
and tests can distinguish the "handler-not-called"
case from the "handler called but wrong value seen"
case.

Test: fx test configuration-tests-component
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/892707
Original-Revision: f6d2468e9c658d2bd75d9cec0ae9e991c579431a
GitOrigin-RevId: 9600c688642583df0d62879ff3d6053335185443
Change-Id: I48569fe1673804cfa4697407f9050a45581cdd26