[roll] Roll fuchsia [testsharder] Check that test deps are built
Currently, it's the recipe that enforces that every file listed in a
shard's deps exists on-disk. However, from the recipe's point of view,
deps are aggregated by shard, so when a dep doesn't exist we can't know
which test the dep belongs to, which would be helpful information for
debugging (as the test likely failed to hook up its deps properly in the
build graph).

Now we'll do that validation in testsharder instead, where we're able to
know which test each dep belongs to and print the test's name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896778
Original-Revision: d4d54db5dfae4b27f79a86b0f1614d500c03bd16
GitOrigin-RevId: 81bb1d8634edb4e164f5430b590a16fc92598925
Change-Id: I77d28d065ff70bef3e67f6ff1adad61da049c4f7