[roll] Roll fuchsia [fxtest][rewrite] Console output module
This CL implements console output for the new fxtest. The output
consists of a few lines of continually updated status strings (if
--status is true and stdout is a TTY). Formatting is controlled by
command line flags passed in to the async task.

The test runs a basic scenario for now, but will be expended to full
golden file tests when the console format is stabilized.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891789
Original-Revision: a57fd255289eaecf0c3f2efc3192dafc903a231d
GitOrigin-RevId: 693b42ec8da34f7dcf1378d5e527d4012fc5a2ee
Change-Id: I46daa03ce349b90bf6d3121cd6211b8fe0c9786f