[roll] Roll fuchsia [ld] Add LdStartupCreateProcessTests
This adds a second test fixture for testing loading the startup
dynamic linker directly in a freshly-created Zircon process.  The
tests approximate what the system program loader does, but with
transparent steps that can be controlled and debugged in the test
code.  This factors out some common code shared with the Zircon
version of the LdStartupInProcessTests fixture.  It also changes
the TestProcessArgs and its uses to be somewhat more generic to
any potential bootstrap protocol.  This looks forward to adding
additional versions of this test fixture that use new bootstrap
protocols as those get developed and prototyped.

The test executables are now compiled with separate entry point
code for the in-process and separate-process cases.  The POSIX
separate-process version is built now, and will be tested later
when a POSIX separate-process test fixture is implemented.  The
Zircon separate-process version isn't quite ready to be used yet
because the test executable requires symbolic relocation for the
zx_process_exit system call from the vDSO and that is not yet
present in the startup dynamic linker.  However, local testing
has verified that the LdStartupInProcessTests case works as far
as launching dynamic linker and the executable correctly up to
the point it tries to make the system call.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895581
Original-Revision: 7fdb0213368ca34bce240d0171023933f2248d32
GitOrigin-RevId: 90ec8fa338dd5c4d02a64301eba41225965dd6c3
Change-Id: Id624cef1eb8ec4c5cc6d54561d6a664bd7d34e0c