[roll] Roll fuchsia [fxfs] Fix flushing a file where last reference is from a VMO
Prior to this CL, Fxfs would panic if the last reference is from a VMO
and the file needed flushing.  The problem was that it was attempting to
spawn an async task from the thread that handles the pager port, which
isn't possible and results in a panic.

I took the opportunity to refactor how the pager works with respect to
spawning tasks and shutdown.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896539
Original-Revision: 1384c3e3ca9f827363ff4aab765877ed5a3b981b
GitOrigin-RevId: b5a82261e7746a48c3620efea5bd66959cea1801
Change-Id: I6a10ba7e4473a5e9395d7175349355f5bbcfe003