[roll] Roll fuchsia [starnix] Make sure hard linked nodes resolve to same node
Hard links need to resolve to the same underlying FsNode.  Also, not all
underlying remote filesystems guarantee to provide unique IDs (and they
could span filesystems anyway), so this changes remote filesystems so
that IDs are only used if the file system is Fxfs, which is known to
have unique IDs.

A previous iteration of this CL started down the path of trying to use a
mount option, but that doesn't work for remote filesystems created via
components where there's no opportunity to specify mount options.
Still, the refactoring I did to support this seems like it might be
beneficial in future so I've left it in.

Original-Bug: 293949025
Original-Bug: 131807
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896521
Original-Revision: 9ae0d3eb502744951ab448e18f999af09cbe2b68
GitOrigin-RevId: bb05d7262f9db301029bf604db8fb0646d4a20a9
Change-Id: Ie1601d4cdcdc7efe637d5c30fa3f7f62e3b01976