[roll] Roll fuchsia [early-boot-instrumentation] Flexible kernel files
No longer hard code the path, instead use a convention where
instrumentation data is under:
 * kernel/i/

Each entry under that path is a directory whose name is the sink name of
the debug data.
 * kernel/i/sink-name/

Each sink directory has up to two subdirectories:
 * kernel/i/sink-name/s/ -> static data
 * kernel/i/sink-name/d/ -> dynamic data.

Each of this directories has a named file representing the module.suffix
the kernel wish to expose. For now we can use the 31 characters
available for the object name to encode this information. If this
becomes a problem it can be revisited. For now, name is just truncated.

Then early boot component exports each of the corresponding files with
the respective file name.

Side-effect filenames are no longer remapped, but that is fine, since
test manager just copies over static/dynamic directories as is.

Original-Bug: 120396 # Done
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822975
Original-Revision: 80c5fa106bdc92dee90623c10f985b3d45518034
GitOrigin-RevId: 8d95db338f1157ab12344f78776b0329d5d1aef0
Change-Id: Iff97713502118666552c6c5ab3efb2b8f704a17a