[roll] Roll fuchsia [netstack3] Set timer while lock held in INCOMPLETE
Currently, when a new neighbor entry is created in the INCOMPLETE state,
the NUD lock is dropped before a neighbor probe is sent and a timer is
scheduled to retransmit a probe. There is a possible race condition
where an INCOMPLETE entry transitions into another state before the
timer is scheduled, at which point it would attempt to cancel a
non-existent timer. Instead, schedule the time while the NUD lock is
held.

To do this, introduce a NudConfigContext trait that allows accessing the
device-specific NUD configuration (retransmit timer in this case) while
holding the NUD lock.

Also add a loom test demonstrating the race condition and fix.

Original-Bug: 35185
Test: netstack3-core-test
      netstack3_core_threading_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897339
Original-Revision: 5e211bdf5553b3a787fae83c6cdaa6852ada5cd1
GitOrigin-RevId: 9bfa6cceea47a41ca3f24aa09c57d3b3ee806691
Change-Id: I548cec12ae6ae1009576b962c77c45b564fe5e6e