[roll] Roll fuchsia [starnix][signals] Prepare for different kinds of waits
Previously, the signal machinery assumed that a task could only sleep in
a Waiter. This CL abstracts this field so that we can sleep in different
ways in the future.  In a subsequent CL, we will introduce more kinds of
sleeps that are optimized for different purposes, with the first one
being optimized for futex waits.

Original-Bug: 131974
Test: No behavior change
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898398
Original-Revision: ed59a0a158040727b02ec060bebd02a56a9e215d
GitOrigin-RevId: cd0d9c584c274eb156bba71b5165257f6d08df14
Change-Id: Id4e69084106127bc09fb47b99e319877ef0f5981