[roll] Roll fuchsia [boot-shim] Introduce SetBootOptionsWithoutEntropy
SetBootOptions when used in the context of shims must hace the option to
not consume the entropy of the command line, since in many cases is the sole
source of entropy and is meant to be used as a seed for kernel's
entropy.

Other boot shims also manually restore the entropy bits when parsing
the serial command line.

Original-Bug: 129401
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895879
Original-Revision: d1f5482baf064ae1f087008a21ab37cb89dfcc92
GitOrigin-RevId: 1b43c07147c91d1c81c50ae0a4001727da5d768d
Change-Id: If9704b7f5286a21d7544ea5129408b64cdf7c355