[roll] Roll fuchsia [netstack3] Implement dual-stack UDP bind()
Implement binding to an IPv4-mapped IPv6 address for UDP. The
implementation for datagram::listen_inner exhaustively matches on the
cases for IP address and dual-stack-supportedness to prove that it
handles all variants. This uses the new dual-stack context trait
introduced in a previous commit to provide access to state.

Original-Bug: 21198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895578
Original-Revision: d212d129a900ac2017050cb01020520fef57c982
GitOrigin-RevId: 9953bbdd87f098db58861710400c8a69b8307304
Change-Id: I722f6e2681302188c739938dd1a6281e2050fa86