[roll] Roll fuchsia [fxfs] Fix fuzzer issue (not panic when close)
The fuzzer catches the panic when we close the filesystem in the fuzzing
test. Remove the panic here. `fs.close()` can fail in a situation where
we have objects to be tombstoned in the journal and that object in the
tree is corrupt. When compaction is called, this will fail (as expected).

Original-Fixed: b/293619473
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899099
Original-Revision: 50f2e4cddd8d48f8553765c7004f27f8c2f1c0de
GitOrigin-RevId: a1d3e81b6e3b2a59d9c5e304431e2c3ea5eac602
Change-Id: I344e06c88b3b5da0f962a69ab7eaaa0ccb310fb7