[roll] Roll fuchsia [fidl] Remove GetMessageBodyTypeNames and simplify fidlgen_go
This CL removes the fidlgen function GetMessageBodyTypeNames. Backends
should not care whether a type is used as a message body. This is the
final step of the refactor planned in https://fxbug.dev/123307.

This mainly involved changing fidlgen_go. Before, it went to great pains
to make requests and responses lowercase (unexported) if the type was
anonymous in the .fidl file, which then required duplicating the type if
it was composed from another library. This is not how it's supposed to
work: in FIDL, an anonymous type should be exactly equivalent to a named
type except fidlc chose the name for you. Fixing this made fidlgen_go a
lot simpler, and as a side effect it also made @generated_name work
properly in Go, fixing https://fxbug.dev/105840.

There are also some changes to syzkaller goldens because it no longer
omits anonymous request/response types. There is no reason to do that.

Doing this revealed an issue with @bindings_denylist and composed
methods (https://fxbug.dev/131989), so I also fixed that in types.go.

Test: fx check-goldens fidl
Original-Fixed: 123307, 131989, 105840
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898427
Original-Revision: c96178fd207cd0ae275b3ec8f2a486ca3d2b3b07
GitOrigin-RevId: 539a1e59917e523e1daae428c0574eecae50872b
Change-Id: I0768087dae2fe47a289fe22f9ebdc6ddbca575ff