[roll] Roll infra/recipes [fxt] Update fxt output format.
cl/553840589 updates fxt to unconditionally output a
`presubmit_token` field, which will always be empty at the moment.
This CL updates the tests to include this field, which is kind of
a "yadda yadda" diff compared to the changes in the next CL I'll be
submitting.

Original-Bug: b/294421350

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/897158
Original-Revision: 5d7362db91de77f0a6ed49749def36a8b4ac181f
GitOrigin-RevId: f0dd25b72a6565e2cd2c1b17461b3c380492f17d
Change-Id: I28997119cf789e4c403920169783a8ea4c649ffb