[roll] Roll fuchsia [flatland][tests] split pixeltests into multiple components
Split out the Flatland pixeltests into two components,
to ensure that the tests don't time out.

We currently have 29 tests, each of which is expected to
take about 10 seconds. That puts us uncomfortably close
to the timeout of 300 seconds.

The margin is low enough to cause flake in practice.

We split the flip-and-orientation tests out into a
separate component, since the various permutations of
pixel format, flip, and orientation, add up to 16
tests.

Original-Bug: 131955
Test: ran on CV, verified all tests still run
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899038
Original-Revision: 6fe2f73ec0ffa49446c095951b680ce881a04580
GitOrigin-RevId: 40fc52aa1d1bc563cfa3fa6b742eb81045ea5b42
Change-Id: I96ec60c240e08cfaf73f5c4cd45c40456de766f2