[roll] Roll fuchsia [time] Adds `SIGNAL_UTC_CLOCK_SYNCHRONIZED`.
When this signal is asserted, the clock was recently synchronized with
an external source reasonably recently. The error bounds represent the
system's best estimate of wall clock UTC time.

When this signal is de-asserted, the clock was last updated from a
source that is not external or was never updated at all. This can occur
after a reboot, or a resume from suspend. The UTC clock may have been
read from the real time clock (RTC), but we do not yet know if RTC
readout matches actual wall clock.

You may wait on this signal if you want your code to sleep until the UTC
clock gets synchronized with an external time source.

If you need finer-grained info on the UTC clock error, you must check
the error bounds via [`zx_clock_get_details`][det].

This signal being asserted does *not* establish clock trust, and should
not be used as such.

The value is an alias for `ZX_USER_SIGNAL_0`.

[det]:
https://fuchsia.dev/fuchsia-src/reference/kernel_objects/clock#error-bound

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898902
Original-Revision: 293f35a286f699c756a859d0dee74f14421c65f1
GitOrigin-RevId: 603cb7026039d57789929c3187097a461006db64
Change-Id: I0b17a77e1ed40f9aaa1fef0ec07a4bb73fb9ca2a