[roll] Roll fuchsia [bazel_sdk] put fuchsia_clang_repository in its own file
This CL paves the way for being able to use the fuchsia_sdk_common rules
in the fuchsia_clang_repository rule. We need to pull this out of the
deps.bzl file because the deps.bzl file will load the logic which
creates the common repository. If we load the clang repository here we
will not be able to use the common repository since it will not have
been loaded yet and will lead to errors.

BUG: b/295358711
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899740
Original-Revision: 590c0e4daef7102c526d3f2712603072c8654b5a
GitOrigin-RevId: 12501d8563115a9a7f6fc2e6a5fd7e79cbb261d0
Change-Id: I54c3eaf1c5e9bc77e5cc52539660801128d876ac