[roll] Roll fuchsia [socket-tun] Want double buffer size for NS2
Netstack 2 doubles the value of SO_SNDBUF and SO_RCVBUF on set, and
returns the doubled value on get. Account for this in the assertions
that the setsockopt calls succeed by doubling the want value only for
NS2 (NS3 does not do this doubling).

Also replicate the clamping behavior in both stacks so that a strict
equality can be asserted.

Test: fx test tun-socket-benchmarks
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/894999
Original-Revision: c93a700f113db30982e360c4dc0dc46c63557e3e
GitOrigin-RevId: 7b6482ed15d699fbdbbb41e1b550d7c36c46efd6
Change-Id: I173d94e03bac3ed6895b260d1aa53d27c09afaca