[roll] Roll fuchsia [wlan] Extend test timeout.
watch_phys tests the ability of the wlandevicemonitor
internals to watch a directory and report the presence of a
new PHY device.  This test relies on driver manager and the
wlantap driver.  It seems like in infra, this test sometimes
takes a while to run, resulting in flakes due to timeouts.
This change extends the timeout to reduce flakiness.

Original-Bug: 131839
Test: fx test wlandevicemonitor-tests
Multiply: wlandevicemonitor-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/897398
Original-Revision: f5f0fb3941ec29dc21a376e0b7852be9ee26bc89
GitOrigin-RevId: e762e563db169faad2fa4c00d3a71de597053a7a
Change-Id: I13deec4b5e6594394e9380dbfc927d96ea2a3b67