[roll] Roll fuchsia [build] Remove unused //build/config/zircon:freestanding
This config is not used at all today, and we don't really want
to use -ffreestanding for compiling even our "standalone" things
like kernel binaries, so we shouldn't need this in the future.

Note the kernel separately using -ffreestanding (but should not)
and this change does not affect that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898558
Original-Revision: d7dd6d2e016a828a99ba3ff0ce219dc043fd8309
GitOrigin-RevId: 39e18680a67c083e334b6b51232301d9afd772d8
Change-Id: I393ac93d3978a6839305920555fa33127a094f51