[roll] Roll fuchsia [ffx][testing] Skip target test if image is too big.
This CL modifies the `ffx target_tests` to disable them if the zircon
image is too big. The test uses gigaboot with UEFI to boot, which places
a hard limit on the zircon kernel image size of 7MiB.

The test is now disabled by build logic if the kernel image is too big.

Note that after CL 874540, physboot will now panic in this same
situation.

Some common reasons for the kernel image to grow too big are compiling
it with extra instrumentation, like `kasan` or `kubsan`.

Original-Bug: 129386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/882094
Original-Revision: 539fbf3a64816f906657d23f110f9b0515e63050
GitOrigin-RevId: 7148222dabd2ad8dd206edfa941b322d7a2d616e
Change-Id: I17fab2ed6bbf33114d9416eefbc69d217f5369b7