[roll] Roll fuchsia [ffx][emu] Adding check for running instance.
This adds a check for a running emulator instance with the same
name. If one is found, an error is returned.

If there is an existing instance directory, but the emulator is not
running, the instance is cleaned up.

Original-Bug: b/295343839
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899547
Original-Revision: 6041a50667ab5dd4c02d9f2b29a5becd35d5f685
GitOrigin-RevId: 3dc2e0401d814fadf61d8cfe39f39f66b1c4cfcf
Change-Id: Ie79a12a0f92a69a8c0d57e4a49ccf3e6e0552d42