[roll] Roll fuchsia [ffx] Remove query from fastboot discovery
The daemon should discover USB devices passively -- there should
be no actual USB traffic until a command requiring the device
is issued, in order to avoid unnecessary crosstalk.  This code
changes the one fastboot check (the validation that the version
of the fastboot device is "0.4") from discovery to when the
device is opened in order to interface (for flashing, rebooting,
etc).

Fixes: b/293909661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898539
Original-Revision: ff69ccb0a601d9fb434bc6797b7fd83d918f7cb3
GitOrigin-RevId: bd2397a74d255e6e7e73fb9a2face3ff108f997b
Change-Id: I94ab56b1629ee017671a82deb0ff8d253f781a70