[roll] Roll fuchsia [build] Fix loadable_module() variants to copy into base root_out_dir
For the output file of a loadable_module() to be the input to
another target such as host_test_data(), there must be a copy
to the known location in the base toolchain's output directory
as is done for executables.  This doesn't come up for the
toolchains that target Fuchsia, since those generally only
consume the output files via metadata anyway.  But it does arise
with the linux_$cpu toolchains that support loadable_module()
targets that need to be used in other ways.

Original-Bug: 130483
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899721
Original-Revision: 82728282d18a281818edd3f37c24aad534735096
GitOrigin-RevId: 073198d3a3182db0e973cf02831b4bc04597b421
Change-Id: I8a7461526918822988737d64bc179b1e361b2a4d