[roll] Roll fuchsia [fuchsia_controller] Send msg, then return coro
This makes it a guarantee that a message will be sent when calling a
FIDL method, even if the coroutine is not awaited. This is also
compatible when interlacing different executors via back-to-back
asyncio.run() calls.

Original-Bug: b/295397172
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899826
Original-Revision: bebc5ed40ee8f2b24b8a625f59a0b0eed9c16dc1
GitOrigin-RevId: e982d332fb4cb63600cd87da69cb4a527433949f
Change-Id: I47cba770c61fe6b784e44070d1a1f8a6d9dd1b2d