[roll] Roll fuchsia [network] Add netstack migration FIDL library
fuchsia.net.stackmigrationdeprecated controls runtime selection of
netstack versions and will be used for migration from Netstack2 to
Netstack3.

It's very targeted and it carries deprecated in its name to communicate
that this is a bespoke set of protocols that should not be copied or
outlive their immediate use.

It's added to the SDK so partners can add switches to start testing
Netstack3 as the main system stack.

Original-Bug: 131101, 131335

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/888311
Original-Revision: 1564a6bd4f5f6e4cbf229a99d7e1947efef529eb
GitOrigin-RevId: 8910524a57534506a3db3a5db378092bfee22c30
Change-Id: Ic913e197b97410385cda19f0bd979c54059bc1a1