[roll] Roll fuchsia [netstack3] Combine datagram socket state and addr
Hold the state and address in the same field. This is a functional no-op
but storing these together will allow them to vary together, which makes
it possible to replace, e.g., ListenerState with an enum that encodes
whether it is bound to allow dual-stack operation.

Original-Bug: 21198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/899723
Original-Revision: e304c6094757774830d7670bbcd63e66d1021d2e
GitOrigin-RevId: a3c381bd59b0250cbeb11ad0086bfebd6872dc56
Change-Id: Ib8272443d3d16fedd67258bfddfad3906933c41a