[roll] Roll fuchsia Reland "[camera-controller] Use FIDL for sysmem protocol"
This is a reland of commit cb7a9987081688316b23de8fb6afee2bf27afc65

Commit was reverted because it was suspected to cause the sysmem-v1 test to fail. This turned out to not be the case.

Original change's description:
> [camera-controller] Use FIDL for sysmem protocol
>
> Make the camera-controller driver use the FIDL version of the
> fuchsia.hardware.gpio protocol instead of the banjo version.
>
> Original-Bug: 131975
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898439

Original-Bug: 131975
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901219
Original-Revision: bf49961fc6d2a8944eac8e928625e62795d64eff
GitOrigin-RevId: 42db95831299c4183b2aff73b6b78cc12275cbfd
Change-Id: Ibcc840db39cc7eb7d018ca1d1a27217363e733ef