[roll] Roll fuchsia Reland "[aml-securemem] Use sysmem fidl protocol"
This is a reland of commit aaee12addf911086cf095e4e484dde1dfe440a27

Commit was reverted because it was suspected to cause the sysmem-v1 test to fail. This turned out to not be the case.

Original change's description:
> [aml-securemem] Use sysmem fidl protocol
>
> Make the `aml-securemem` driver use the FIDL version of the sysmem
> protocol instead of the banjo version.
>
> Original-Bug: 131694
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/896658

Original-Bug: 131694
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901218
Original-Revision: 1aee0e7e4104cdc50191a41ac09b4ee0feffe0d4
GitOrigin-RevId: 97bf81a09e01460e3aace5891dd55cf945c4bfd4
Change-Id: I7a3c791520f2e872db31b72b7b1d5612570721fa